### 2021-01-25 #2986 @whedon
[PRE REVIEW]: covid19.analytics: An R Package to Obtain, Analyze and Visualize Data from the 2019 Corona Virus Disease Pandemic
**Submitting author:** @mponce0 (<a href="http://orcid.org/0000-0001-5850-7240">Marcelo Ponce</a>)
**Repository:** <a href="https://github.com/mponce0/covid19.analytics/" target ="_blank">https://github.com/mponce0/covid19.analytics/</a>
**Version:** v2.0
