### 2021-01-25 #2986 @arfon
Cloc output from local machine:

```
arfon@mawr ~ $ cloc covid19.analytics/
      70 text files.
      64 unique files.
      40 files ignored.

Language                     files          blank        comment           code
R                               21           1186           2098           3738
HTML                             5            438              0           3429
Markdown                         1            295              0           1099
YAML                             2             19             21             24
Bourne Shell                     1              8              6             18
Rmd                              1              3             14              4
SUM:                            31           1949           2139           8312
```
