### 2021-01-25 #2986 @danielskatz
👋  @mponce0 - if you can remove the entries in the .bib file that are not actually being used, that will help reduce the potential problems that check references found.
