+++ JOSS/references.bib
@@ -25,6 +25,7 @@
 
 @Manual{cran,
   title = {{The Comprehensive R Archive Network}},
+  Doi = {0.1002/wics.1212},
   url = {https://cran.r-project.org}
 }
 
@@ -102,6 +103,7 @@
     year = {2020},
     isbn = {9781138331457},
     url = {https://plotly-r.com},
+    Doi = {10.1201/9780429447273}
 }
 
 @Manual{DT,
@@ -129,6 +131,7 @@
     volume = {9},
     issue = {2},
     pages = {520--526},
+    Doi = {10.7287/peerj.preprints.3179v2}
 }
 
 
@@ -139,6 +142,7 @@
     year = {2019},
     volume = {35},
     pages = {526-528},
+    Doi = {10.1093/bioinformatics/bty633}
 }
 
 
@@ -333,7 +337,7 @@
 	Author = {Tiberiu Harko and Francisco S.N. Lobo and M.K. Mak},
 	Date-Added = {2020-07-03 17:15:11 -0400},
 	Date-Modified = {2020-07-03 17:16:01 -0400},
-	Doi = {https://doi.org/10.1016/j.amc.2014.03.030},
+	Doi = {10.1016/j.amc.2014.03.030},
 	Issn = {0096-3003},
 	Journal = {Applied Mathematics and Computation},
 	Keywords = {c19_SIR -- Susceptible-Infected-Recovered (SIR) epidemic model, Exact solution, Abel equation},
@@ -759,7 +763,8 @@
 	Date-Modified = {2020-07-31 10:22:43 -0400},
 	Journal = {PNAS; Proceedings of the National Academy of Sciences},
 	Title = {Timing social distancing to avert unmanageable COVID-19 hospital surges},
-	Year = {2020}
+	Year = {2020},
+	Doi = {10.1073/pnas.2009033117}
 }
 
 
@@ -854,7 +859,7 @@ title = "Tracking Changes in SARS-CoV-2 Spike: Evidence that D614G Increases Inf
 journal = "Cell",
 year = "2020",
 issn = "0092-8674",
-doi = "https://doi.org/10.1016/j.cell.2020.06.043",
+doi = "10.1016/j.cell.2020.06.043",
 url = "http://www.sciencedirect.com/science/article/pii/S0092867420308205",
 author = "Bette Korber and Will M. Fischer and Sandrasegaram Gnanakaran and Hyejin Yoon and James Theiler and Werner Abfalterer and Nick Hengartner and Elena E. Giorgi and Tanmoy Bhattacharya and Brian Foley and Kathryn M. Hastie and Matthew D. Parker and David G. Partridge and Cariad M. Evans and Timothy M. Freeman and Thushan I. {de Silva} and Adrienne Angyal and Rebecca L. Brown and Laura Carrilero and Luke R. Green and Danielle C. Groves and Katie J. Johnson and Alexander J. Keeley and Benjamin B. Lindsey and Paul J. Parsons and Mohammad Raza and Sarah Rowland-Jones and Nikki Smith and Rachel M. Tucker and Dennis Wang and Matthew D. Wyles and Charlene McDanal and Lautaro G. Perez and Haili Tang and Alex Moon-Walker and Sean P. Whelan and Celia C. LaBranche and Erica O. Saphire and David C. Montefiori",
 keywords = "COVID-19, SARS-CoV-2, diversity, evolution, Spike, antibody, infectivity, neutralization, pseudovirus, PCR cycle threshold",
