+++ JOSS/references.bib
@@ -31,17 +31,6 @@
 
 %%%%%%%%%%%%%%%%
 
-@article{gentleman2004bioconductor,
-  title={{Bioconductor: open software development for computational biology and bioinformatics}},
-  author={Gentleman, Robert C and Carey, Vincent J and Bates, Douglas M and Bolstad, Ben and Dettling, Marcel and Dudoit, Sandrine and Ellis, Byron and Gautier, Laurent and Ge, Yongchao and Gentry, Jeff and others},
-  journal={Genome biology},
-  volume={5},
-  number={10},
-  pages={R80},
-  year={2004},
-  publisher={Springer},
-  url={http://www.bioconductor.org}
-}
 
 
 %%%%%%%%%%%% R Packages %%%%%%%%%%%
@@ -64,102 +53,9 @@
 }
 
 
-@Manual{devtools,
-    title = {devtools: Tools to Make Developing R Packages Easier},
-    author = {Hadley Wickham and Jim Hester and Winston Chang},
-    year = {2020},
-    note = {R package version 2.3.0},
-    url = {https://CRAN.R-project.org/package=devtools},
-}
-
-@Manual{shinyPckg,
-    title = {shiny: Web Application Framework for R},
-    author = {Winston Chang and Joe Cheng and JJ Allaire and Yihui Xie and Jonathan McPherson},
-    year = {2020},
-    note = {R package version 1.4.0.2},
-    url = {https://CRAN.R-project.org/package=shiny},
-}
-
-@Manual{shinydashboard,
-    title = {shinydashboard: Create Dashboards with 'Shiny'},
-    author = {Winston Chang and Barbara {Borges Ribeiro}},
-    year = {2018},
-    note = {R package version 0.7.1},
-    url = {https://CRAN.R-project.org/package=shinydashboard},
-}
-
-@Manual{shinycssloaders,
-    title = {shinycssloaders: Add CSS Loading Animations to 'shiny' Outputs},
-    author = {Andras Sali and Dean Attali},
-    year = {2020},
-    note = {R package version 0.3},
-    url = {https://CRAN.R-project.org/package=shinycssloaders},
-}
-
-@Book{plotly,
-    author = {Carson Sievert},
-    title = {Interactive Web-Based Data Visualization with R, plotly, and shiny},
-    publisher = {Chapman and Hall/CRC},
-    year = {2020},
-    isbn = {9781138331457},
-    url = {https://plotly-r.com},
-    Doi = {10.1201/9780429447273}
-}
-
-@Manual{DT,
-    title = {DT: A Wrapper of the JavaScript Library 'DataTables'},
-    author = {Yihui Xie and Joe Cheng and Xianying Tan},
-    year = {2020},
-    note = {R package version 0.13},
-    url = {https://CRAN.R-project.org/package=DT},
-}
-
-@Manual{dplyr,
-    title = {dplyr: A Grammar of Data Manipulation},
-    author = {Hadley Wickham and Romain François and Lionel Henry and Kirill Müller},
-    year = {2020},
-    note = {R package version 1.0.0},
-    url = {https://CRAN.R-project.org/package=dplyr},
-}
-
-
-@Article{rentrezPckg,
-    title = {{rentrez}: an R package for the NCBI eUtils API},
-    author = {David J. Winter},
-    journal = {The R Journal},
-    year = {2017},
-    volume = {9},
-    issue = {2},
-    pages = {520--526},
-    Doi = {10.7287/peerj.preprints.3179v2}
-}
-
-
-@Article{apePckg,
[…]
