+++ JOSS/references.bib
@@ -25,7 +25,7 @@
 
 @Manual{cran,
   title = {{The Comprehensive R Archive Network}},
-  Doi = {0.1002/wics.1212},
+  Doi = {10.1002/wics.1212},
   url = {https://cran.r-project.org}
 }
 
