### 2021-01-26 #2986 @mponce0
> 👋 @mponce0 - if you can remove the entries in the .bib file that are not actually being used, that will help reduce the potential problems that check references found.

@danielskatz Thanks! I fixed all the issues with the DOIs and removed unused items from the .bib file
