### 2021-01-26 b300ae6
daily update of covid19 data/plots/figures: covid19-SummaryReport.txt ... Tue Jan 26 01:03:46 EST 2021-gargantua
