+++ man/figures/covid19-SummaryReport.txt
@@ -1,30 +1,30 @@
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ 
 -------------------------------------------------------------------------------- 
 ################################################################################ 
-  ##### TS-CONFIRMED Cases  -- Data dated:  2021-01-24  ::  2021-01-25 13:02:38 
+  ##### TS-CONFIRMED Cases  -- Data dated:  2021-01-25  ::  2021-01-26 01:02:39 
 ################################################################################ 
   Number of Countries/Regions reported:  192 
   Number of Cities/Provinces reported:  85 
   Unique number of distinct geographical locations combined: 273 
 -------------------------------------------------------------------------------- 
-  Worldwide ts-confirmed  Totals: 99194405 
+  Worldwide ts-confirmed  Totals: 99718414 
 -------------------------------------------------------------------------------- 
    Country.Region Province.State   Totals GlobalPerc LastDayChange    t-2    t-3    t-7   t-14   t-30
-1              US                25127000      25.33        132537 172650 189925 141999 213304 224761
-2           India                10667736      10.75         13203  14849  14256  10050  12584  18732
-3          Brazil                 8844577       8.92         28323  62334  56552  23671  25822  17246
-4          Russia                 3679247       3.71         20800  20585  21182  22509  23018  28833
-5  United Kingdom                 3647463       3.68         30004  33552  40261  37535  46169  34693
-6          France                 3044358       3.07         18038  23410  22995   3736   3582   3030
-7           Spain                 2499560       2.52             0      0  42885  84287  61422      0
-8           Italy                 2466813       2.49         11628  13331  13633   8825  12530  10405
-9          Turkey                 2429605       2.45          5277   5856   5967   5862  10220  15118
-10        Germany                 2147769       2.17         10078  12430  16366   9253  11706  13504
-11       Colombia                 2015485       2.03         12516  15551  15073  14719  15003  10196
-12      Argentina                 1867223       1.88          5031   8362  10753   8185   8704   3713
-13         Mexico                 1763219       1.78         10872  20057  21007   8074   7594   4974
-14         Poland                 1475445       1.49          4566   6431   6693   3332   4863   4878
-15   South Africa                 1412986       1.42          8147  12271  11761   9010  15046  11552
+1              US                25293201      25.36        147254 131191 170156 176992 226150 155690
+2           India                10676838      10.71          9102  13203  14849  13816  15968  20021
+3          Brazil                 8871393       8.90         26816  28323  62334  62094  64025  18479
+4          Russia                 3698246       3.71         18999  20800  20585  21442  22657  27849
+5  United Kingdom                 3669658       3.68         22195  30004  33552  33355  45533  32340
+6          France                 3048598       3.06          4240  18038  23410  23339  19267   8233
+7           Spain                 2593382       2.60         93822      0      0  34291  25438      0
+8           Italy                 2475372       2.48          8559  11628  13331  10496  14242   8937
+9          Turkey                 2435247       2.44          5642   5277   5856   6818   9809  14205
+10        Germany                 2154656       2.16          6887  10078  12430  12233  27210  12399
+11       Colombia                 2027746       2.03         12261  12516  15551  15939  14179   9594
+12      Argentina                 1874801       1.88          7578   5031   8362  12141  13783   5030
+13         Mexico                 1771740       1.78          8521  10872  20057  18894  14395   6217
+14         Poland                 1478119       1.48          2674   4566   6431   4890   5394   3842
+15   South Africa                 1417537       1.42          4551   8147  12271   9780  13105   9502
 -------------------------------------------------------------------------------- 
[…]
