+++ R/pandemics.R
@@ -4,15 +4,18 @@
 
 pandemics.loaddata <- function(tgt.file="pandemics.RDS",
 			acknowledge=TRUE, show=FALSE,
-			src='') {
+			src.descr='') {
 #' internal function to retrieve historical data on pandemics
 #'
 #' @param  tgt.file  which data set to read
 #' @param  acknowledge  displays details on the data sources
 #' @param  show  displays data
-#' @param  src
+#' @param  src.descr  description of the source of the data
 #'
 #' @return data.frame
+#'
+#'
+#' @importFrom  DT  datatable
 #'
 
 	covid19.pckg <- 'covid19.analytics'
@@ -21,7 +24,7 @@ pandemics.loaddata <- function(tgt.file="pandemics.RDS",
 
 	if (file.exists(pnds.RDS)) {
 		pnd.data <- readRDS(pnds.RDS)
-		if (acknowledge) message("Data obtained from",src)
+		if (acknowledge) message("Data obtained from",src.descr)
 		if (show) print(datatable(pnd.data))
 		return(pnd.data)
 	} else {
@@ -41,14 +44,14 @@ pandemics.data <- function(acknowledge=TRUE, show=FALSE, tgt='pandemics') {
 #'
 
 	if (tgt=='pandemics') {
-		src <- "Pandemic historical records -- data from  https://www.visualcapitalist.com/history-of-pandemics-deadliest/"
+		src.descr <- "Pandemic historical records -- data from  https://www.visualcapitalist.com/history-of-pandemics-deadliest/"
 		tgt <- 'pandemics.RDS'
 	} else if (tgt=='pandemics_vaccines') {
-		src <- "Pandemics vaccine development times -- data from  https://www.visualcapitalist.com/the-race-to-save-lives-comparing-vaccine-development-timelines/"
+		src.descr <- "Pandemics vaccine development times -- data from  https://www.visualcapitalist.com/the-race-to-save-lives-comparing-vaccine-development-timelines/"
 		tgt <- 'pandemics_vaccines.RDS'
 	} else {
 		stop("Unrecognized option! Possible options for tgt are 'pandemics' OR 'pandemics_vaccines'")
 	}
 
-	return(pandemics.loaddata(tgt, acknowledge, show, src))
+	return(pandemics.loaddata(tgt, acknowledge, show, src.descr))
 }
