### 2021-01-26 61ab800
daily update of covid19 data/plots/figures: covid19-SummaryReport.txt ... Tue Jan 26 13:03:47 EST 2021-gargantua
