+++ man/figures/covid19-SummaryReport.txt
@@ -1,16 +1,16 @@
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ 
 -------------------------------------------------------------------------------- 
 ################################################################################ 
-  ##### TS-CONFIRMED Cases  -- Data dated:  2021-01-25  ::  2021-01-26 01:02:39 
+  ##### TS-CONFIRMED Cases  -- Data dated:  2021-01-25  ::  2021-01-26 13:02:39 
 ################################################################################ 
   Number of Countries/Regions reported:  192 
   Number of Cities/Provinces reported:  85 
   Unique number of distinct geographical locations combined: 273 
 -------------------------------------------------------------------------------- 
-  Worldwide ts-confirmed  Totals: 99718414 
+  Worldwide ts-confirmed  Totals: 99722272 
 -------------------------------------------------------------------------------- 
    Country.Region Province.State   Totals GlobalPerc LastDayChange    t-2    t-3    t-7   t-14   t-30
-1              US                25293201      25.36        147254 131191 170156 176992 226150 155690
+1              US                25297059      25.37        151112 131191 170156 176992 226150 155690
 2           India                10676838      10.71          9102  13203  14849  13816  15968  20021
 3          Brazil                 8871393       8.90         26816  28323  62334  62094  64025  18479
 4          Russia                 3698246       3.71         18999  20800  20585  21442  22657  27849
@@ -33,16 +33,16 @@
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ 
 -------------------------------------------------------------------------------- 
 ################################################################################ 
-  ##### TS-DEATHS Cases  -- Data dated:  2021-01-25  ::  2021-01-26 01:02:40 
+  ##### TS-DEATHS Cases  -- Data dated:  2021-01-25  ::  2021-01-26 13:02:40 
 ################################################################################ 
   Number of Countries/Regions reported:  192 
   Number of Cities/Provinces reported:  85 
   Unique number of distinct geographical locations combined: 273 
 -------------------------------------------------------------------------------- 
-  Worldwide ts-deaths  Totals: 2139928 
+  Worldwide ts-deaths  Totals: 2140085 
 -------------------------------------------------------------------------------- 
    Country.Region Province.State Totals Perc LastDayChange  t-2  t-3  t-7 t-14 t-30
-1              US                420972 1.66          1758 1775 3332 2769 4462 1214
+1              US                421129 1.66          1915 1775 3332 2769 4462 1214
 2          Brazil                217664 2.45           627  592 1202 1192 1110  344
 3           India                153587 1.44           117  131  155  162  202  279
 4          Mexico                150273 8.48           659  530 1470 1584 1314  400
@@ -63,7 +63,7 @@
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ 
 -------------------------------------------------------------------------------- 
 ################################################################################ 
-  ##### TS-RECOVERED Cases  -- Data dated:  2021-01-25  ::  2021-01-26 01:02:40 
+  ##### TS-RECOVERED Cases  -- Data dated:  2021-01-25  ::  2021-01-26 13:02:40 
 ################################################################################ 
   Number of Countries/Regions reported:  192 
   Number of Cities/Provinces reported:  69 
[…]
