### 2021-01-27 8c6e08e
daily update of covid19 data/plots/figures: covid19-SummaryReport.txt ... Wed Jan 27 01:03:39 EST 2021-gargantua
