+++ man/figures/covid19-SummaryReport.txt
@@ -1,30 +1,30 @@
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ 
 -------------------------------------------------------------------------------- 
 ################################################################################ 
-  ##### TS-CONFIRMED Cases  -- Data dated:  2021-01-25  ::  2021-01-26 13:02:39 
+  ##### TS-CONFIRMED Cases  -- Data dated:  2021-01-26  ::  2021-01-27 01:02:38 
 ################################################################################ 
   Number of Countries/Regions reported:  192 
   Number of Cities/Provinces reported:  85 
   Unique number of distinct geographical locations combined: 273 
 -------------------------------------------------------------------------------- 
-  Worldwide ts-confirmed  Totals: 99722272 
+  Worldwide ts-confirmed  Totals: 100270535 
 -------------------------------------------------------------------------------- 
    Country.Region Province.State   Totals GlobalPerc LastDayChange    t-2    t-3    t-7   t-14   t-30
-1              US                25297059      25.37        151112 131191 170156 176992 226150 155690
-2           India                10676838      10.71          9102  13203  14849  13816  15968  20021
-3          Brazil                 8871393       8.90         26816  28323  62334  62094  64025  18479
-4          Russia                 3698246       3.71         18999  20800  20585  21442  22657  27849
-5  United Kingdom                 3669658       3.68         22195  30004  33552  33355  45533  32340
-6          France                 3048598       3.06          4240  18038  23410  23339  19267   8233
-7           Spain                 2593382       2.60         93822      0      0  34291  25438      0
-8           Italy                 2475372       2.48          8559  11628  13331  10496  14242   8937
-9          Turkey                 2435247       2.44          5642   5277   5856   6818   9809  14205
-10        Germany                 2154656       2.16          6887  10078  12430  12233  27210  12399
-11       Colombia                 2027746       2.03         12261  12516  15551  15939  14179   9594
-12      Argentina                 1874801       1.88          7578   5031   8362  12141  13783   5030
-13         Mexico                 1771740       1.78          8521  10872  20057  18894  14395   6217
-14         Poland                 1478119       1.48          2674   4566   6431   4890   5394   3842
-15   South Africa                 1417537       1.42          4551   8147  12271   9780  13105   9502
+1              US                25439570      25.37        142511 151112 131191 182532 229786 174090
+2           India                10689527      10.66         12689   9102  13203  15244  16946  16432
+3          Brazil                 8933356       8.91         61963  26816  28323  64385  60899  20548
+4          Russia                 3716228       3.71         17982  18999  20800  20806  22544  27363
+5  United Kingdom                 3689746       3.68         20088  22195  30004  38905  47525  41385
+6          France                 3070458       3.06         21860   4240  18038  26497  23703   2954
+7           Spain                 2629817       2.62         36435  93822      0  41576  38869  24462
+8           Italy                 2485956       2.48         10584   8559  11628  13568  15773   8581
+9          Turkey                 2442350       2.44          7103   5642   5277   6435   9554  15197
+10        Germany                 2164043       2.16          9387   6887  10078  29003  25566  14004
+11       Colombia                 2041352       2.04         13606  12261  12516  17908  15898   9310
+12      Argentina                 1885210       1.88         10409   7578   5031  12112  12725   7216
+13         Mexico                 1788905       1.78         17165   8521  10872  20548  15873   5996
+14         Poland                 1482722       1.48          4603   2674   4566   6943   9126   3211
+15   South Africa                 1423578       1.42          6041   4551   8147  12710  18555   7458
 -------------------------------------------------------------------------------- 
[…]
