### 2021-01-27 b496950
daily update of covid19 data/plots/figures: covid19-SummaryReport.txt ... Wed Jan 27 13:03:44 EST 2021-gargantua
