+++ man/figures/covid19-SummaryReport.txt
@@ -1,16 +1,16 @@
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ 
 -------------------------------------------------------------------------------- 
 ################################################################################ 
-  ##### TS-CONFIRMED Cases  -- Data dated:  2021-01-26  ::  2021-01-27 01:02:38 
+  ##### TS-CONFIRMED Cases  -- Data dated:  2021-01-26  ::  2021-01-27 13:02:40 
 ################################################################################ 
   Number of Countries/Regions reported:  192 
   Number of Cities/Provinces reported:  85 
   Unique number of distinct geographical locations combined: 273 
 -------------------------------------------------------------------------------- 
-  Worldwide ts-confirmed  Totals: 100270535 
+  Worldwide ts-confirmed  Totals: 100274664 
 -------------------------------------------------------------------------------- 
    Country.Region Province.State   Totals GlobalPerc LastDayChange    t-2    t-3    t-7   t-14   t-30
-1              US                25439570      25.37        142511 151112 131191 182532 229786 174090
+1              US                25443699      25.37        146640 151112 131191 182532 229786 174090
 2           India                10689527      10.66         12689   9102  13203  15244  16946  16432
 3          Brazil                 8933356       8.91         61963  26816  28323  64385  60899  20548
 4          Russia                 3716228       3.71         17982  18999  20800  20806  22544  27363
@@ -33,16 +33,16 @@
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ 
 -------------------------------------------------------------------------------- 
 ################################################################################ 
-  ##### TS-DEATHS Cases  -- Data dated:  2021-01-26  ::  2021-01-27 01:02:39 
+  ##### TS-DEATHS Cases  -- Data dated:  2021-01-26  ::  2021-01-27 13:02:41 
 ################################################################################ 
   Number of Countries/Regions reported:  192 
   Number of Cities/Provinces reported:  85 
   Unique number of distinct geographical locations combined: 273 
 -------------------------------------------------------------------------------- 
-  Worldwide ts-deaths  Totals: 2157349 
+  Worldwide ts-deaths  Totals: 2157446 
 -------------------------------------------------------------------------------- 
    Country.Region Province.State Totals Perc LastDayChange  t-2  t-3  t-7 t-14 t-30
-1              US                425119 1.67          3990 1915 1775 4375 3963 1841
+1              US                425216 1.67          4087 1915 1775 4375 3963 1841
 2          Brazil                218878 2.45          1214  627  592 1340 1274  431
 3           India                153724 1.44           137  117  131  151  198  252
 4          Mexico                152016 8.50          1743  659  530 1539 1235  429
@@ -63,7 +63,7 @@
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ 
 -------------------------------------------------------------------------------- 
 ################################################################################ 
-  ##### TS-RECOVERED Cases  -- Data dated:  2021-01-26  ::  2021-01-27 01:02:39 
+  ##### TS-RECOVERED Cases  -- Data dated:  2021-01-26  ::  2021-01-27 13:02:42 
 ################################################################################ 
   Number of Countries/Regions reported:  192 
[…]
