### 2021-01-28 8bec076
daily update of covid19 data/plots/figures: covid19-SummaryReport.txt ... Thu Jan 28 01:03:40 EST 2021-gargantua
