+++ man/figures/covid19-SummaryReport.txt
@@ -1,30 +1,30 @@
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ 
 -------------------------------------------------------------------------------- 
 ################################################################################ 
-  ##### TS-CONFIRMED Cases  -- Data dated:  2021-01-26  ::  2021-01-27 13:02:40 
+  ##### TS-CONFIRMED Cases  -- Data dated:  2021-01-27  ::  2021-01-28 01:02:37 
 ################################################################################ 
   Number of Countries/Regions reported:  192 
   Number of Cities/Provinces reported:  85 
   Unique number of distinct geographical locations combined: 273 
 -------------------------------------------------------------------------------- 
-  Worldwide ts-confirmed  Totals: 100274664 
+  Worldwide ts-confirmed  Totals: 100866032 
 -------------------------------------------------------------------------------- 
    Country.Region Province.State   Totals GlobalPerc LastDayChange    t-2    t-3    t-7   t-14   t-30
-1              US                25443699      25.37        146640 151112 131191 182532 229786 174090
-2           India                10689527      10.66         12689   9102  13203  15244  16946  16432
-3          Brazil                 8933356       8.91         61963  26816  28323  64385  60899  20548
-4          Russia                 3716228       3.71         17982  18999  20800  20806  22544  27363
-5  United Kingdom                 3689746       3.68         20088  22195  30004  38905  47525  41385
-6          France                 3070458       3.06         21860   4240  18038  26497  23703   2954
-7           Spain                 2629817       2.62         36435  93822      0  41576  38869  24462
-8           Italy                 2485956       2.48         10584   8559  11628  13568  15773   8581
-9          Turkey                 2442350       2.44          7103   5642   5277   6435   9554  15197
-10        Germany                 2164043       2.16          9387   6887  10078  29003  25566  14004
-11       Colombia                 2041352       2.04         13606  12261  12516  17908  15898   9310
-12      Argentina                 1885210       1.88         10409   7578   5031  12112  12725   7216
-13         Mexico                 1788905       1.78         17165   8521  10872  20548  15873   5996
-14         Poland                 1482722       1.48          4603   2674   4566   6943   9126   3211
-15   South Africa                 1423578       1.42          6041   4551   8147  12710  18555   7458
+1              US                25598061      25.38        152478 146597 151095 193055 234975 198880
+2           India                10701193      10.61         11666  12689   9102  14545  15590  20549
+3          Brazil                 8996876       8.92         63520  61963  26816  59119  67758  58718
+4          Russia                 3733692       3.70         17464  17982  18999  21544  24303  26588
+5  United Kingdom                 3715054       3.68         25308  20088  22195  37892  48682  53135
+6          France                 3097374       3.07         26916  21860   4240  22368  21228  11192
+7           Spain                 2670102       2.65         40285  36435  93822  44357  35878  14089
+8           Italy                 2501147       2.48         15191  10584   8559  14055  17243  11210
+9          Turkey                 2449839       2.43          7489   7103   5642   6289   8962  15805
+10        Germany                 2179679       2.16         15636   9387   6887   8277  21343  19466
+11       Colombia                 2055305       2.04         13953  13606  12261  15366  17121  11015
+12      Argentina                 1896053       1.88         10843  10409   7578  11396  13286  11650
+13         Mexico                 1806849       1.79         17944  17165   8521  22339  16468  12099
+14         Poland                 1489512       1.48          6790   4603   2674   7008   9457   7624
+15   South Africa                 1430648       1.42          7070   6041   4551  11381  18503   9580
 -------------------------------------------------------------------------------- 
[…]
