### 2021-01-28 bb2ee57
daily update of covid19 data/plots/figures: covid19-SummaryReport.txt ... Thu Jan 28 13:03:40 EST 2021-gargantua
