+++ man/figures/covid19-SummaryReport.txt
@@ -1,7 +1,7 @@
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ 
 -------------------------------------------------------------------------------- 
 ################################################################################ 
-  ##### TS-CONFIRMED Cases  -- Data dated:  2021-01-27  ::  2021-01-28 01:02:37 
+  ##### TS-CONFIRMED Cases  -- Data dated:  2021-01-27  ::  2021-01-28 13:02:39 
 ################################################################################ 
   Number of Countries/Regions reported:  192 
   Number of Cities/Provinces reported:  85 
@@ -33,7 +33,7 @@
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ 
 -------------------------------------------------------------------------------- 
 ################################################################################ 
-  ##### TS-DEATHS Cases  -- Data dated:  2021-01-27  ::  2021-01-28 01:02:38 
+  ##### TS-DEATHS Cases  -- Data dated:  2021-01-27  ::  2021-01-28 13:02:39 
 ################################################################################ 
   Number of Countries/Regions reported:  192 
   Number of Cities/Provinces reported:  85 
@@ -63,7 +63,7 @@
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ 
 -------------------------------------------------------------------------------- 
 ################################################################################ 
-  ##### TS-RECOVERED Cases  -- Data dated:  2021-01-27  ::  2021-01-28 01:02:38 
+  ##### TS-RECOVERED Cases  -- Data dated:  2021-01-27  ::  2021-01-28 13:02:40 
 ################################################################################ 
   Number of Countries/Regions reported:  192 
   Number of Cities/Provinces reported:  69 
@@ -92,7 +92,7 @@
 ================================================================================ 
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ 
 ############################################################################################################################################ 
-  ##### AGGREGATED Data  -- ORDERED BY  CONFIRMED Cases  -- Data dated:  2021-01-28  ::  2021-01-28 01:02:38 
+  ##### AGGREGATED Data  -- ORDERED BY  CONFIRMED Cases  -- Data dated:  2021-01-28  ::  2021-01-28 13:02:40 
 ############################################################################################################################################ 
   Number of Countries/Regions reported: 192 
   Number of Cities/Provinces reported: 576 
@@ -116,7 +116,7 @@
 15               Kerala, India    905591           0.90   3663        0.40    829452          91.59   72476        8.00
 ============================================================================================================================================ 
 ############################################################################################################################################ 
-  ##### AGGREGATED Data  -- ORDERED BY  DEATHS Cases  -- Data dated:  2021-01-28  ::  2021-01-28 01:02:38 
+  ##### AGGREGATED Data  -- ORDERED BY  DEATHS Cases  -- Data dated:  2021-01-28  ::  2021-01-28 13:02:40 
 ############################################################################################################################################ 
   Number of Countries/Regions reported: 192 
   Number of Cities/Provinces reported: 576 
@@ -140,7 +140,7 @@
 15                  Romania    718612           0.71  18015        2.51    662400          92.18   38197        5.32
 ============================================================================================================================================ 
 ############################################################################################################################################ 
[…]
