### 2021-01-29 1640a96
daily update of covid19 data/plots/figures: covid19-SummaryReport.txt ... Fri Jan 29 01:04:13 EST 2021-gargantua
