+++ man/figures/covid19-SummaryReport.txt
@@ -1,30 +1,30 @@
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ 
 -------------------------------------------------------------------------------- 
 ################################################################################ 
-  ##### TS-CONFIRMED Cases  -- Data dated:  2021-01-27  ::  2021-01-28 13:02:39 
+  ##### TS-CONFIRMED Cases  -- Data dated:  2021-01-28  ::  2021-01-29 01:02:37 
 ################################################################################ 
   Number of Countries/Regions reported:  192 
   Number of Cities/Provinces reported:  85 
   Unique number of distinct geographical locations combined: 273 
 -------------------------------------------------------------------------------- 
-  Worldwide ts-confirmed  Totals: 100866032 
+  Worldwide ts-confirmed  Totals: 101453597 
 -------------------------------------------------------------------------------- 
    Country.Region Province.State   Totals GlobalPerc LastDayChange    t-2    t-3    t-7   t-14   t-30
-1              US                25598061      25.38        152478 146597 151095 193055 234975 198880
-2           India                10701193      10.61         11666  12689   9102  14545  15590  20549
-3          Brazil                 8996876       8.92         63520  61963  26816  59119  67758  58718
-4          Russia                 3733692       3.70         17464  17982  18999  21544  24303  26588
-5  United Kingdom                 3715054       3.68         25308  20088  22195  37892  48682  53135
-6          France                 3097374       3.07         26916  21860   4240  22368  21228  11192
-7           Spain                 2670102       2.65         40285  36435  93822  44357  35878  14089
-8           Italy                 2501147       2.48         15191  10584   8559  14055  17243  11210
-9          Turkey                 2449839       2.43          7489   7103   5642   6289   8962  15805
-10        Germany                 2179679       2.16         15636   9387   6887   8277  21343  19466
-11       Colombia                 2055305       2.04         13953  13606  12261  15366  17121  11015
-12      Argentina                 1896053       1.88         10843  10409   7578  11396  13286  11650
-13         Mexico                 1806849       1.79         17944  17165   8521  22339  16468  12099
-14         Poland                 1489512       1.48          6790   4603   2674   7008   9457   7624
-15   South Africa                 1430648       1.42          7070   6041   4551  11381  18503   9580
+1              US                25762726      25.39        164665 152478 146597 190032 240707 232016
+2           India                10720048      10.57         18855  11666  12689  14256  15158  21822
+3          Brazil                 9058687       8.93         61811  63520  61963  56552  69198  55649
+4          Russia                 3752548       3.70         18856  17464  17982  21182  24294  26095
+5  United Kingdom                 3743734       3.69         28680  25308  20088  40261  55761  50023
+6          France                 3097374       3.05             0  26916  21860  22995  21068  26357
+7           Spain                 2705001       2.67         34899  40285  36435  42885  40197  16716
+8           Italy                 2515507       2.48         14360  15191  10584  13633  16144  16202
+9          Turkey                 2457118       2.42          7279   7489   7103   5967   8314  15692
+10        Germany                 2194562       2.16         14883  15636   9387  16366   8593  49044
+11       Colombia                 2067575       2.04         12270  13953  13606  15073  21078  11639
+12      Argentina                 1905524       1.88          9471  10843  10409  10753  12332  11765
+13         Mexico                 1825519       1.80         18670  17944  17165  21007  21366  12406
+14         Poland                 1496665       1.48          7153   6790   4603   6693   7958  12780
+15   South Africa                 1437798       1.42          7150   7070   6041  11761  14880  17710
 -------------------------------------------------------------------------------- 
[…]
