### 2021-01-29 93f5bfb
daily update of covid19 data/plots/figures: covid19-SummaryReport.txt ... Fri Jan 29 13:03:48 EST 2021-gargantua
