+++ man/figures/covid19-SummaryReport.txt
@@ -1,16 +1,16 @@
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ 
 -------------------------------------------------------------------------------- 
 ################################################################################ 
-  ##### TS-CONFIRMED Cases  -- Data dated:  2021-01-28  ::  2021-01-29 01:02:37 
+  ##### TS-CONFIRMED Cases  -- Data dated:  2021-01-28  ::  2021-01-29 13:02:39 
 ################################################################################ 
   Number of Countries/Regions reported:  192 
   Number of Cities/Provinces reported:  85 
   Unique number of distinct geographical locations combined: 273 
 -------------------------------------------------------------------------------- 
-  Worldwide ts-confirmed  Totals: 101453597 
+  Worldwide ts-confirmed  Totals: 101457552 
 -------------------------------------------------------------------------------- 
    Country.Region Province.State   Totals GlobalPerc LastDayChange    t-2    t-3    t-7   t-14   t-30
-1              US                25762726      25.39        164665 152478 146597 190032 240707 232016
+1              US                25766681      25.40        168620 152478 146597 190032 240707 232016
 2           India                10720048      10.57         18855  11666  12689  14256  15158  21822
 3          Brazil                 9058687       8.93         61811  63520  61963  56552  69198  55649
 4          Russia                 3752548       3.70         18856  17464  17982  21182  24294  26095
@@ -33,16 +33,16 @@
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ 
 -------------------------------------------------------------------------------- 
 ################################################################################ 
-  ##### TS-DEATHS Cases  -- Data dated:  2021-01-28  ::  2021-01-29 01:02:38 
+  ##### TS-DEATHS Cases  -- Data dated:  2021-01-28  ::  2021-01-29 13:02:39 
 ################################################################################ 
   Number of Countries/Regions reported:  192 
   Number of Cities/Provinces reported:  85 
   Unique number of distinct geographical locations combined: 273 
 -------------------------------------------------------------------------------- 
-  Worldwide ts-deaths  Totals: 2191078 
+  Worldwide ts-deaths  Totals: 2191206 
 -------------------------------------------------------------------------------- 
    Country.Region Province.State Totals Perc LastDayChange  t-2  t-3  t-7 t-14 t-30
-1              US                433067 1.68          3872 3943 4084 3760 3805 3749
+1              US                433195 1.68          4000 3943 4084 3760 3805 3749
 2          Brazil                221547 2.45          1386 1283 1214 1096 1151 1194
 3          Mexico                155145 8.50          1506 1623 1743 1440 1106 1052
 4           India                154010 1.44           163  123  137  152  175  299
@@ -63,7 +63,7 @@
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ 
 -------------------------------------------------------------------------------- 
 ################################################################################ 
-  ##### TS-RECOVERED Cases  -- Data dated:  2021-01-28  ::  2021-01-29 01:02:38 
+  ##### TS-RECOVERED Cases  -- Data dated:  2021-01-28  ::  2021-01-29 13:02:40 
 ################################################################################ 
[…]
