### 2021-01-30 cce9edf
daily update of covid19 data/plots/figures: covid19-SummaryReport.txt ... Sat Jan 30 01:03:36 EST 2021-gargantua
