+++ man/figures/covid19-SummaryReport.txt
@@ -1,30 +1,30 @@
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ 
 -------------------------------------------------------------------------------- 
 ################################################################################ 
-  ##### TS-CONFIRMED Cases  -- Data dated:  2021-01-28  ::  2021-01-29 13:02:39 
+  ##### TS-CONFIRMED Cases  -- Data dated:  2021-01-29  ::  2021-01-30 01:02:36 
 ################################################################################ 
   Number of Countries/Regions reported:  192 
   Number of Cities/Provinces reported:  85 
   Unique number of distinct geographical locations combined: 273 
 -------------------------------------------------------------------------------- 
-  Worldwide ts-confirmed  Totals: 101457552 
+  Worldwide ts-confirmed  Totals: 102065711 
 -------------------------------------------------------------------------------- 
    Country.Region Province.State   Totals GlobalPerc LastDayChange    t-2    t-3    t-7   t-14   t-30
-1              US                25766681      25.40        168620 152478 146597 190032 240707 232016
-2           India                10720048      10.57         18855  11666  12689  14256  15158  21822
-3          Brazil                 9058687       8.93         61811  63520  61963  56552  69198  55649
-4          Russia                 3752548       3.70         18856  17464  17982  21182  24294  26095
-5  United Kingdom                 3743734       3.69         28680  25308  20088  40261  55761  50023
-6          France                 3097374       3.05             0  26916  21860  22995  21068  26357
-7           Spain                 2705001       2.67         34899  40285  36435  42885  40197  16716
-8           Italy                 2515507       2.48         14360  15191  10584  13633  16144  16202
-9          Turkey                 2457118       2.42          7279   7489   7103   5967   8314  15692
-10        Germany                 2194562       2.16         14883  15636   9387  16366   8593  49044
-11       Colombia                 2067575       2.04         12270  13953  13606  15073  21078  11639
-12      Argentina                 1905524       1.88          9471  10843  10409  10753  12332  11765
-13         Mexico                 1825519       1.80         18670  17944  17165  21007  21366  12406
-14         Poland                 1496665       1.48          7153   6790   4603   6693   7958  12780
-15   South Africa                 1437798       1.42          7150   7070   6041  11761  14880  17710
+1              US                25929282      25.40        162601 168620 152478 170138 200764 233916
+2           India                10733130      10.52         13082  18855  11666  14849  15144      0
+3          Brazil                 9118513       8.93         59826  61811  63520  62334  61567  56773
+4  United Kingdom                 3772813       3.70         29079  28680  25308  33552  41342  55892
+5          Russia                 3771514       3.70         18966  18856  17464  20585  23670  27329
+6          France                 3143417       3.08         46043      0  26916  23410      0  19778
+7           Spain                 2743119       2.69         38118  34899  40285      0      0  18047
+8           Italy                 2529070       2.48         13563  14360  15191  13331  16310  23477
+9          Turkey                 2464030       2.41          6912   7279   7489   5856   7550  14380
+10        Germany                 2207393       2.16         12831  14883  15636  12430  14817  19367
+11       Colombia                 2077633       2.04         10058  12270  13953  15551  20855  16314
+12      Argentina                 1915362       1.88          9838   9471  10843   8362   8932  11586
+13         Mexico                 1841893       1.80         16374  18670  17944  20057  20523  12159
+14         Poland                 1502810       1.47          6145   7153   6790   6431   7292  13464
+15   South Africa                 1443939       1.41          6141   7150   7070  12271  13973  18000
[…]
