### 2021-01-30 f5af36b
daily update of covid19 data/plots/figures: covid19-SummaryReport.txt ... Sat Jan 30 13:03:37 EST 2021-gargantua
