+++ man/figures/covid19-SummaryReport.txt
@@ -1,16 +1,16 @@
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ 
 -------------------------------------------------------------------------------- 
 ################################################################################ 
-  ##### TS-CONFIRMED Cases  -- Data dated:  2021-01-29  ::  2021-01-30 01:02:36 
+  ##### TS-CONFIRMED Cases  -- Data dated:  2021-01-29  ::  2021-01-30 13:02:37 
 ################################################################################ 
   Number of Countries/Regions reported:  192 
   Number of Cities/Provinces reported:  85 
   Unique number of distinct geographical locations combined: 273 
 -------------------------------------------------------------------------------- 
-  Worldwide ts-confirmed  Totals: 102065711 
+  Worldwide ts-confirmed  Totals: 102069223 
 -------------------------------------------------------------------------------- 
    Country.Region Province.State   Totals GlobalPerc LastDayChange    t-2    t-3    t-7   t-14   t-30
-1              US                25929282      25.40        162601 168620 152478 170138 200764 233916
+1              US                25932794      25.41        166113 168620 152478 170138 200764 233916
 2           India                10733130      10.52         13082  18855  11666  14849  15144      0
 3          Brazil                 9118513       8.93         59826  61811  63520  62334  61567  56773
 4  United Kingdom                 3772813       3.70         29079  28680  25308  33552  41342  55892
@@ -33,16 +33,16 @@
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ 
 -------------------------------------------------------------------------------- 
 ################################################################################ 
-  ##### TS-DEATHS Cases  -- Data dated:  2021-01-29  ::  2021-01-30 01:02:37 
+  ##### TS-DEATHS Cases  -- Data dated:  2021-01-29  ::  2021-01-30 13:02:38 
 ################################################################################ 
   Number of Countries/Regions reported:  192 
   Number of Cities/Provinces reported:  85 
   Unique number of distinct geographical locations combined: 273 
 -------------------------------------------------------------------------------- 
-  Worldwide ts-deaths  Totals: 2206336 
+  Worldwide ts-deaths  Totals: 2206457 
 -------------------------------------------------------------------------------- 
    Country.Region Province.State Totals Perc LastDayChange  t-2  t-3  t-7 t-14 t-30
-1              US                436678 1.68          3483 4000 3943 3329 3352 3413
+1              US                436799 1.68          3604 4000 3943 3329 3352 3413
 2          Brazil                222666 2.44          1119 1386 1283 1202 1050 1074
 3          Mexico                156579 8.50          1434 1506 1623 1470 1219  910
 4           India                154147 1.44           137  163  123  155  181    0
@@ -63,7 +63,7 @@
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ 
 -------------------------------------------------------------------------------- 
 ################################################################################ 
-  ##### TS-RECOVERED Cases  -- Data dated:  2021-01-29  ::  2021-01-30 01:02:38 
+  ##### TS-RECOVERED Cases  -- Data dated:  2021-01-29  ::  2021-01-30 13:02:38 
 ################################################################################ 
   Number of Countries/Regions reported:  192 
   Number of Cities/Provinces reported:  69 
@@ -92,7 +92,7 @@
[…]
