### 2021-01-31 36eaf06
daily update of covid19 data/plots/figures: covid19-SummaryReport.txt ... Sun Jan 31 01:03:26 EST 2021-gargantua
