+++ man/figures/covid19-SummaryReport.txt
@@ -1,30 +1,30 @@
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ 
 -------------------------------------------------------------------------------- 
 ################################################################################ 
-  ##### TS-CONFIRMED Cases  -- Data dated:  2021-01-29  ::  2021-01-30 13:02:37 
+  ##### TS-CONFIRMED Cases  -- Data dated:  2021-01-30  ::  2021-01-31 01:02:35 
 ################################################################################ 
   Number of Countries/Regions reported:  192 
   Number of Cities/Provinces reported:  85 
   Unique number of distinct geographical locations combined: 273 
 -------------------------------------------------------------------------------- 
-  Worldwide ts-confirmed  Totals: 102069223 
+  Worldwide ts-confirmed  Totals: 102577980 
 -------------------------------------------------------------------------------- 
    Country.Region Province.State   Totals GlobalPerc LastDayChange    t-2    t-3    t-7   t-14   t-30
-1              US                25932794      25.41        166113 168620 152478 170138 200764 233916
-2           India                10733130      10.52         13082  18855  11666  14849  15144      0
-3          Brazil                 9118513       8.93         59826  61811  63520  62334  61567  56773
-4  United Kingdom                 3772813       3.70         29079  28680  25308  33552  41342  55892
-5          Russia                 3771514       3.70         18966  18856  17464  20585  23670  27329
-6          France                 3143417       3.08         46043      0  26916  23410      0  19778
-7           Spain                 2743119       2.69         38118  34899  40285      0      0  18047
-8           Italy                 2529070       2.48         13563  14360  15191  13331  16310  23477
-9          Turkey                 2464030       2.41          6912   7279   7489   5856   7550  14380
-10        Germany                 2207393       2.16         12831  14883  15636  12430  14817  19367
-11       Colombia                 2077633       2.04         10058  12270  13953  15551  20855  16314
-12      Argentina                 1915362       1.88          9838   9471  10843   8362   8932  11586
-13         Mexico                 1841893       1.80         16374  18670  17944  20057  20523  12159
-14         Poland                 1502810       1.47          6145   7153   6790   6431   7292  13464
-15   South Africa                 1443939       1.41          6141   7150   7070  12271  13973  18000
+1              US                26069046      25.41        136252 166113 168620 131075 177404 152344
+2           India                10746174      10.48         13044  13082  18855  13203  13788  20035
+3          Brazil                 9176975       8.95         58462  59826  61811  28323  33040  24605
+4  United Kingdom                 3796088       3.70         23275  29079  28680  30004  38598  53285
+5          Russia                 3790265       3.70         18751  18966  18856  20800  23178  26613
+6          France                 3167274       3.09         23857  46043      0  18038  37206  19143
+7           Spain                 2743119       2.67             0  38118  34899      0      0      0
+8           Italy                 2541783       2.48         12713  13563  14360  11628  12544  22210
+9          Turkey                 2470901       2.41          6871   6912   7279   5277   6436  12203
+10        Germany                 2224911       2.17         17518  12831  14883  10078  11484   2117
+11       Colombia                 2086806       2.03          9173  10058  12270  12516  17379  12105
+12      Argentina                 1922264       1.87          6902   9838   9471   5031   7264   4080
+13         Mexico                 1857230       1.81         15337  16374  18670  10872  11170  11091
+14         Poland                 1508674       1.47          5864   6145   7153   4566   5970  10896
+15   South Africa                 1449236       1.41          5297   6141   7150   8147  12267  16726
 -------------------------------------------------------------------------------- 
[…]
