### 2021-01-31 41abe3c
daily update of covid19 data/plots/figures: covid19-SummaryReport.txt ... Sun Jan 31 13:03:35 EST 2021-gargantua
