+++ man/figures/covid19-SummaryReport.txt
@@ -1,20 +1,20 @@
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ 
 -------------------------------------------------------------------------------- 
 ################################################################################ 
-  ##### TS-CONFIRMED Cases  -- Data dated:  2021-01-30  ::  2021-01-31 01:02:35 
+  ##### TS-CONFIRMED Cases  -- Data dated:  2021-01-30  ::  2021-01-31 13:02:37 
 ################################################################################ 
   Number of Countries/Regions reported:  192 
   Number of Cities/Provinces reported:  85 
   Unique number of distinct geographical locations combined: 273 
 -------------------------------------------------------------------------------- 
-  Worldwide ts-confirmed  Totals: 102577980 
+  Worldwide ts-confirmed  Totals: 102583819 
 -------------------------------------------------------------------------------- 
    Country.Region Province.State   Totals GlobalPerc LastDayChange    t-2    t-3    t-7   t-14   t-30
-1              US                26069046      25.41        136252 166113 168620 131075 177404 152344
+1              US                26074885      25.42        142091 166113 168620 131075 177404 152344
 2           India                10746174      10.48         13044  13082  18855  13203  13788  20035
 3          Brazil                 9176975       8.95         58462  59826  61811  28323  33040  24605
 4  United Kingdom                 3796088       3.70         23275  29079  28680  30004  38598  53285
-5          Russia                 3790265       3.70         18751  18966  18856  20800  23178  26613
+5          Russia                 3790265       3.69         18751  18966  18856  20800  23178  26613
 6          France                 3167274       3.09         23857  46043      0  18038  37206  19143
 7           Spain                 2743119       2.67             0  38118  34899      0      0      0
 8           Italy                 2541783       2.48         12713  13563  14360  11628  12544  22210
@@ -33,16 +33,16 @@
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ 
 -------------------------------------------------------------------------------- 
 ################################################################################ 
-  ##### TS-DEATHS Cases  -- Data dated:  2021-01-30  ::  2021-01-31 01:02:36 
+  ##### TS-DEATHS Cases  -- Data dated:  2021-01-30  ::  2021-01-31 13:02:38 
 ################################################################################ 
   Number of Countries/Regions reported:  192 
   Number of Cities/Provinces reported:  85 
   Unique number of distinct geographical locations combined: 273 
 -------------------------------------------------------------------------------- 
-  Worldwide ts-deaths  Totals: 2219889 
+  Worldwide ts-deaths  Totals: 2219980 
 -------------------------------------------------------------------------------- 
    Country.Region Province.State Totals Perc LastDayChange  t-2  t-3  t-7 t-14 t-30
-1              US                439439 1.69          2640 3604 4000 1775 1751 2027
+1              US                439530 1.69          2731 3604 4000 1775 1751 2027
 2          Brazil                223945 2.44          1279 1119 1386  592  551  462
 3          Mexico                158074 8.51          1495 1434 1506  530  463  700
 4           India                154274 1.44           127  137  163  131  145  256
@@ -63,7 +63,7 @@
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ 
[…]
