### 2021-02-01 f03e391
daily update of covid19 data/plots/figures: covid19-SummaryReport.txt ... Mon Feb  1 01:03:26 EST 2021-gargantua
