+++ man/figures/covid19-SummaryReport.txt
@@ -1,30 +1,30 @@
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ 
 -------------------------------------------------------------------------------- 
 ################################################################################ 
-  ##### TS-CONFIRMED Cases  -- Data dated:  2021-01-30  ::  2021-01-31 13:02:37 
+  ##### TS-CONFIRMED Cases  -- Data dated:  2021-01-31  ::  2021-02-01 01:02:35 
 ################################################################################ 
   Number of Countries/Regions reported:  192 
   Number of Cities/Provinces reported:  85 
   Unique number of distinct geographical locations combined: 273 
 -------------------------------------------------------------------------------- 
-  Worldwide ts-confirmed  Totals: 102583819 
+  Worldwide ts-confirmed  Totals: 102964429 
 -------------------------------------------------------------------------------- 
    Country.Region Province.State   Totals GlobalPerc LastDayChange    t-2    t-3    t-7   t-14   t-30
-1              US                26074885      25.42        142091 166113 168620 131075 177404 152344
-2           India                10746174      10.48         13044  13082  18855  13203  13788  20035
-3          Brazil                 9176975       8.95         58462  59826  61811  28323  33040  24605
-4  United Kingdom                 3796088       3.70         23275  29079  28680  30004  38598  53285
-5          Russia                 3790265       3.69         18751  18966  18856  20800  23178  26613
-6          France                 3167274       3.09         23857  46043      0  18038  37206  19143
-7           Spain                 2743119       2.67             0  38118  34899      0      0      0
-8           Italy                 2541783       2.48         12713  13563  14360  11628  12544  22210
-9          Turkey                 2470901       2.41          6871   6912   7279   5277   6436  12203
-10        Germany                 2224911       2.17         17518  12831  14883  10078  11484   2117
-11       Colombia                 2086806       2.03          9173  10058  12270  12516  17379  12105
-12      Argentina                 1922264       1.87          6902   9838   9471   5031   7264   4080
-13         Mexico                 1857230       1.81         15337  16374  18670  10872  11170  11091
-14         Poland                 1508674       1.47          5864   6145   7153   4566   5970  10896
-15   South Africa                 1449236       1.41          5297   6141   7150   8147  12267  16726
+1              US                26185355      25.43        110470 142091 166113 151095 142605 300282
+2           India                10757610      10.45         11436  13044  13082   9102  10050  37256
+3          Brazil                 9204731       8.94         27756  58462  59826  26816  23671  15827
+4  United Kingdom                 3817176       3.71         21088  23275  29079  22195  37535  57724
+5          Russia                 3808348       3.70         18083  18751  18966  18999  22509  25938
+6          France                 3186354       3.09         19080  23857  46043   4240   3736   3359
+7           Spain                 2743119       2.66             0      0  38118  93822  84287      0
+8           Italy                 2553032       2.48         11249  12713  13563   8559   8825  11825
+9          Turkey                 2477463       2.41          6562   6871   6912   5642   5862  11180
+10        Germany                 2225659       2.16           748  17518  12831   6887   9253  10903
+11       Colombia                 2094884       2.03          8078   9173  10058  12261  14719  11528
+12      Argentina                 1927239       1.87          4975   6902   9838   7578   8185   5240
+13         Mexico                 1864260       1.81          7030  15337  16374   8521   8074   6359
+14         Poland                 1513385       1.47          4711   5864   6145   2674   3332   7006
+15   South Africa                 1453761       1.41          4525   5297   6141   4551   9010  15002
 -------------------------------------------------------------------------------- 
[…]
