### 2021-02-01 cec6a1a
daily update of covid19 data/plots/figures: covid19-SummaryReport.txt ... Mon Feb  1 13:03:35 EST 2021-gargantua
