+++ man/figures/covid19-SummaryReport.txt
@@ -1,16 +1,16 @@
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ 
 -------------------------------------------------------------------------------- 
 ################################################################################ 
-  ##### TS-CONFIRMED Cases  -- Data dated:  2021-01-31  ::  2021-02-01 01:02:35 
+  ##### TS-CONFIRMED Cases  -- Data dated:  2021-01-31  ::  2021-02-01 13:02:38 
 ################################################################################ 
   Number of Countries/Regions reported:  192 
   Number of Cities/Provinces reported:  85 
   Unique number of distinct geographical locations combined: 273 
 -------------------------------------------------------------------------------- 
-  Worldwide ts-confirmed  Totals: 102964429 
+  Worldwide ts-confirmed  Totals: 102965855 
 -------------------------------------------------------------------------------- 
    Country.Region Province.State   Totals GlobalPerc LastDayChange    t-2    t-3    t-7   t-14   t-30
-1              US                26185355      25.43        110470 142091 166113 151095 142605 300282
+1              US                26186781      25.43        111896 142091 166113 151095 142605 300282
 2           India                10757610      10.45         11436  13044  13082   9102  10050  37256
 3          Brazil                 9204731       8.94         27756  58462  59826  26816  23671  15827
 4  United Kingdom                 3817176       3.71         21088  23275  29079  22195  37535  57724
@@ -33,16 +33,16 @@
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ 
 -------------------------------------------------------------------------------- 
 ################################################################################ 
-  ##### TS-DEATHS Cases  -- Data dated:  2021-01-31  ::  2021-02-01 01:02:36 
+  ##### TS-DEATHS Cases  -- Data dated:  2021-01-31  ::  2021-02-01 13:02:38 
 ################################################################################ 
   Number of Countries/Regions reported:  192 
   Number of Cities/Provinces reported:  85 
   Unique number of distinct geographical locations combined: 273 
 -------------------------------------------------------------------------------- 
-  Worldwide ts-deaths  Totals: 2227900 
+  Worldwide ts-deaths  Totals: 2227905 
 -------------------------------------------------------------------------------- 
    Country.Region Province.State Totals Perc LastDayChange  t-2  t-3  t-7 t-14 t-30
-1              US                441319 1.69          1789 2731 3604 1917 1405 2328
+1              US                441324 1.69          1794 2731 3604 1917 1405 2328
 2          Brazil                224504 2.44           559 1279 1119  627  452  314
 3          Mexico                158536 8.50           462 1495 1434  659  544  344
 4           India                154392 1.44           118  127  137  117  137  441
@@ -63,7 +63,7 @@
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ 
 -------------------------------------------------------------------------------- 
 ################################################################################ 
-  ##### TS-RECOVERED Cases  -- Data dated:  2021-01-31  ::  2021-02-01 01:02:36 
+  ##### TS-RECOVERED Cases  -- Data dated:  2021-01-31  ::  2021-02-01 13:02:39 
 ################################################################################ 
   Number of Countries/Regions reported:  192 
   Number of Cities/Provinces reported:  69 
@@ -92,7 +92,7 @@
[…]
