### 2021-02-02 c3d10f6
daily update of covid19 data/plots/figures: covid19-SummaryReport.txt ... Tue Feb  2 01:03:26 EST 2021-gargantua
