+++ man/figures/covid19-SummaryReport.txt
@@ -1,30 +1,30 @@
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ 
 -------------------------------------------------------------------------------- 
 ################################################################################ 
-  ##### TS-CONFIRMED Cases  -- Data dated:  2021-01-31  ::  2021-02-01 13:02:38 
+  ##### TS-CONFIRMED Cases  -- Data dated:  2021-02-01  ::  2021-02-02 01:02:35 
 ################################################################################ 
   Number of Countries/Regions reported:  192 
   Number of Cities/Provinces reported:  85 
   Unique number of distinct geographical locations combined: 273 
 -------------------------------------------------------------------------------- 
-  Worldwide ts-confirmed  Totals: 102965855 
+  Worldwide ts-confirmed  Totals: 103409402 
 -------------------------------------------------------------------------------- 
    Country.Region Province.State   Totals GlobalPerc LastDayChange    t-2    t-3    t-7   t-14   t-30
-1              US                26186781      25.43        111896 142091 166113 151095 142605 300282
-2           India                10757610      10.45         11436  13044  13082   9102  10050  37256
-3          Brazil                 9204731       8.94         27756  58462  59826  26816  23671  15827
-4  United Kingdom                 3817176       3.71         21088  23275  29079  22195  37535  57724
-5          Russia                 3808348       3.70         18083  18751  18966  18999  22509  25938
-6          France                 3186354       3.09         19080  23857  46043   4240   3736   3359
-7           Spain                 2743119       2.66             0      0  38118  93822  84287      0
-8           Italy                 2553032       2.48         11249  12713  13563   8559   8825  11825
-9          Turkey                 2477463       2.41          6562   6871   6912   5642   5862  11180
-10        Germany                 2225659       2.16           748  17518  12831   6887   9253  10903
-11       Colombia                 2094884       2.03          8078   9173  10058  12261  14719  11528
-12      Argentina                 1927239       1.87          4975   6902   9838   7578   8185   5240
-13         Mexico                 1864260       1.81          7030  15337  16374   8521   8074   6359
-14         Poland                 1513385       1.47          4711   5864   6145   2674   3332   7006
-15   South Africa                 1453761       1.41          4525   5297   6141   4551   9010  15002
+1              US                26317540      25.45        130759 111896 142091 146597 176984 208556
+2           India                10766245      10.41          8635  11436  13044  12689  13816  16504
+3          Brazil                 9229322       8.93         24591  27756  58462  61963  62094  17341
+4  United Kingdom                 3835783       3.71         18607  21088  23275  20088  33355  54990
+5          Russia                 3825739       3.70         17391  18083  18751  17982  21442  23845
+6          France                 3190701       3.09          4347  19080  23857  21860  23339  12489
+7           Spain                 2822805       2.73         79686      0      0  36435  34291      0
+8           Italy                 2560957       2.48          7925  11249  12713  10584  10496  14245
+9          Turkey                 2485182       2.40          7719   6562   6871   7103   6818   9877
+10        Germany                 2232327       2.16          6668    748  17518   9387  12233  10356
+11       Colombia                 2104506       2.04          9622   8078   9173  13606  15939   9412
+12      Argentina                 1933853       1.87          6614   4975   6902  10409  12141   5884
+13         Mexico                 1869708       1.81          5448   7030  15337  17165  18894   5211
+14         Poland                 1515889       1.47          2504   4711   5864   4603   4890   5782
+15   South Africa                 1456309       1.41          2548   4525   5297   6041   9780  11859
 -------------------------------------------------------------------------------- 
[…]
