+++ NAMESPACE
@@ -32,6 +32,7 @@ export(mtrends)
 export(nullify.data)
 export(plt.SIR.model)
 export(preProcessingData)
+export(report.Tor)
 export(report.summary)
 export(rollingRate)
 export(single.trend)
@@ -39,6 +40,7 @@ export(sweep.SIR.models)
 export(totals.plt)
 export(tots.per.location)
 importFrom(DT,dataTableOutput)
+importFrom(DT,datatable)
 importFrom(DT,renderDataTable)
 importFrom(ape,read.FASTA)
 importFrom(ape,read.GenBank)
@@ -52,6 +54,7 @@ importFrom(dplyr,first)
 importFrom(dplyr,mutate)
 importFrom(gplots,bluered)
 importFrom(gplots,heatmap.2)
+importFrom(grDevices,dev.new)
 importFrom(grDevices,heat.colors)
 importFrom(grDevices,rainbow)
 importFrom(grDevices,rgb)
