### 2021-02-02 7eec063
updating data acq script
