### 2021-02-02 16fa1fe
reading local latest agg data from symlink
