### 2021-02-02 d911159
daily update of covid19 data/plots/figures: covid19-SummaryReport.txt ... Tue Feb  2 13:03:45 EST 2021-gargantua
