+++ man/figures/covid19-SummaryReport.txt
@@ -1,18 +1,18 @@
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ 
 -------------------------------------------------------------------------------- 
 ################################################################################ 
-  ##### TS-CONFIRMED Cases  -- Data dated:  2021-02-01  ::  2021-02-02 01:02:35 
+  ##### TS-CONFIRMED Cases  -- Data dated:  2021-02-01  ::  2021-02-02 13:02:38 
 ################################################################################ 
   Number of Countries/Regions reported:  192 
   Number of Cities/Provinces reported:  85 
   Unique number of distinct geographical locations combined: 273 
 -------------------------------------------------------------------------------- 
-  Worldwide ts-confirmed  Totals: 103409402 
+  Worldwide ts-confirmed  Totals: 103412982 
 -------------------------------------------------------------------------------- 
    Country.Region Province.State   Totals GlobalPerc LastDayChange    t-2    t-3    t-7   t-14   t-30
-1              US                26317540      25.45        130759 111896 142091 146597 176984 208556
+1              US                26321120      25.45        134339 111896 142091 146597 176984 208556
 2           India                10766245      10.41          8635  11436  13044  12689  13816  16504
-3          Brazil                 9229322       8.93         24591  27756  58462  61963  62094  17341
+3          Brazil                 9229322       8.92         24591  27756  58462  61963  62094  17341
 4  United Kingdom                 3835783       3.71         18607  21088  23275  20088  33355  54990
 5          Russia                 3825739       3.70         17391  18083  18751  17982  21442  23845
 6          France                 3190701       3.09          4347  19080  23857  21860  23339  12489
@@ -33,16 +33,16 @@
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ 
 -------------------------------------------------------------------------------- 
 ################################################################################ 
-  ##### TS-DEATHS Cases  -- Data dated:  2021-02-01  ::  2021-02-02 01:02:36 
+  ##### TS-DEATHS Cases  -- Data dated:  2021-02-01  ::  2021-02-02 13:02:39 
 ################################################################################ 
   Number of Countries/Regions reported:  192 
   Number of Cities/Provinces reported:  85 
   Unique number of distinct geographical locations combined: 273 
 -------------------------------------------------------------------------------- 
-  Worldwide ts-deaths  Totals: 2237973 
+  Worldwide ts-deaths  Totals: 2238123 
 -------------------------------------------------------------------------------- 
    Country.Region Province.State Totals Perc LastDayChange  t-2  t-3  t-7 t-14 t-30
-1              US                443205 1.68          1881 1794 2731 4084 2774 1344
+1              US                443355 1.68          2031 1794 2731 4084 2774 1344
 2          Brazil                225099 2.44           595  559 1279 1214 1192  293
 3          Mexico                159100 8.51           564  462 1495 1743 1584  362
 4           India                154486 1.43            94  118  127  137  162  214
@@ -63,7 +63,7 @@
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ 
 -------------------------------------------------------------------------------- 
 ################################################################################ 
-  ##### TS-RECOVERED Cases  -- Data dated:  2021-02-01  ::  2021-02-02 01:02:36 
[…]
