+++ R/covid19_vaccination_testing.R
@@ -1,5 +1,5 @@
-# functions to obtain covid19 vaccination data
-# SRC: Our World In Data
+# functions to obtain covid19 Vaccination & Testing data
+# SRC: "Our World In Data" (OWID)
 #
 # M.Ponce - covid19.analytics 
 
@@ -7,36 +7,46 @@
 OWID.repos <- function(tgt) {
 #' function to define the OWID repos URLs
 #'
-#' @param  tgt  target case:  'global','us','country','locations'
+#' @param  tgt  target case:  'VAC.global','VAC.us','VAC.country','VAC.locations', 'testing'
 #'
 #' @return  URL
 #'
 
 	# OWID gtihub repo
-	owid.repo <- "https://raw.githubusercontent.com/owid/covid-19-data/master/public/data/vaccinations/"
-	vacc.locns <- paste0(owid.repo,"locations.csv")
+	#owid.repo <- "https://raw.githubusercontent.com/owid/covid-19-data/master/public/data/vaccinations/"
+	owid.repo <- "https://raw.githubusercontent.com/owid/covid-19-data/master/public/data/"
+	vacc.repo <- paste0(owid.repo,"vaccinations/")
+	vacc.locns <- paste0(vacc.repo,"locations.csv")
+	testing.repo <-  paste0(owid.repo,"testing/")
+	testing.data <- paste0(testing.repo, "covid-testing-all-observations.csv")
+	testing.data.details <- paste0(testing.repo,"covid-testing-latest-data-source-details.csv")
 
 	# United States vaccination data
 	#
 	# State-by-state data on United States COVID-19 vaccinations. We rely on the data updated daily by the United States Centers for Disease Control and Prevention. Stored in us_state_vaccinations.csv.
-	vacc.us <- paste0(owid.repo,"us_state_vaccinations.csv")
+	vacc.us <- paste0(vacc.repo,"us_state_vaccinations.csv")
 
 	# Vaccination data
 	#
 	# Country-by-country data on global COVID-19 vaccinations. We only rely on figures that are verifiable based on public official sources. Stored in vaccinations.csv.
-	vacc.global <- paste0(owid.repo,"vaccinations.csv")
+	vacc.global <- paste0(vacc.repo,"vaccinations.csv")
 
 	# specific per country -- it will be one file per country
-	vacc.country <- paste0(owid.repo,"country_data/")
+	vacc.country <- paste0(vacc.repo,"country_data/")
 
-	if (tgt == 'locations') {
+	# VACINATION REPOS
+	if (tgt == 'VAC.locations') {
 		repo <- vacc.locns
-	} else if (tgt == 'us') {
+	} else if (tgt == 'VAC.us') {
 		repo <- vacc.us
-	} else if (tgt == 'global') {
+	} else if (tgt == 'VAC.global') {
 		repo <- vacc.global
-	} else if (tgt == 'country') {
+	} else if (tgt == 'VAC.country') {
 		repo <- vacc.country
+	# TESTING REPOS
+	} else if (tgt == 'testing') {
+		repo <- testing.data
+	# UNRECOGNIZED option
 	} else {
 		stop("Error : unrecognized target! -- ",tgt)
 	}
@@ -45,11 +55,14 @@ OWID.repos <- function(tgt) {
 }
 
 
-c19.vaccination <- function(repo, disclaimer=TRUE){
-#' function to read data from OWID repo
+c19.OWID.data <- function(repo, disclaimer=TRUE){
+#' function to read data from OWID repos
 #'
 #' @param  repo  URL for reading the data
 #' @param  disclaimer  indicate whether the information about the source of the data is disclosed 
+#'
+#' @return  datafram object
+#'
 
[…]
