+++ NAMESPACE
@@ -17,6 +17,8 @@ export(covid19.URL_csv.data)
 export(covid19.US.data)
 export(covid19.data)
 export(covid19.genomic.data)
+export(covid19.testing.data)
+export(covid19.vaccination)
 export(covid19Explorer)
 export(covid19dashboard)
 export(data.checks)
