### 2021-02-03 d00fd89
daily update of covid19 data/plots/figures: covid19-SummaryReport.txt ... Wed Feb  3 01:03:26 EST 2021-gargantua
