+++ man/figures/covid19-SummaryReport.txt
@@ -1,30 +1,30 @@
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ 
 -------------------------------------------------------------------------------- 
 ################################################################################ 
-  ##### TS-CONFIRMED Cases  -- Data dated:  2021-02-01  ::  2021-02-02 13:02:38 
+  ##### TS-CONFIRMED Cases  -- Data dated:  2021-02-02  ::  2021-02-03 01:02:36 
 ################################################################################ 
   Number of Countries/Regions reported:  192 
   Number of Cities/Provinces reported:  85 
   Unique number of distinct geographical locations combined: 273 
 -------------------------------------------------------------------------------- 
-  Worldwide ts-confirmed  Totals: 103412982 
+  Worldwide ts-confirmed  Totals: 103866068 
 -------------------------------------------------------------------------------- 
    Country.Region Province.State   Totals GlobalPerc LastDayChange    t-2    t-3    t-7   t-14   t-30
-1              US                26321120      25.45        134339 111896 142091 146597 176984 208556
-2           India                10766245      10.41          8635  11436  13044  12689  13816  16504
-3          Brazil                 9229322       8.92         24591  27756  58462  61963  62094  17341
-4  United Kingdom                 3835783       3.71         18607  21088  23275  20088  33355  54990
-5          Russia                 3825739       3.70         17391  18083  18751  17982  21442  23845
-6          France                 3190701       3.09          4347  19080  23857  21860  23339  12489
-7           Spain                 2822805       2.73         79686      0      0  36435  34291      0
-8           Italy                 2560957       2.48          7925  11249  12713  10584  10496  14245
-9          Turkey                 2485182       2.40          7719   6562   6871   7103   6818   9877
-10        Germany                 2232327       2.16          6668    748  17518   9387  12233  10356
-11       Colombia                 2104506       2.04          9622   8078   9173  13606  15939   9412
-12      Argentina                 1933853       1.87          6614   4975   6902  10409  12141   5884
-13         Mexico                 1869708       1.81          5448   7030  15337  17165  18894   5211
-14         Poland                 1515889       1.47          2504   4711   5864   4603   4890   5782
-15   South Africa                 1456309       1.41          2548   4525   5297   6041   9780  11859
+1              US                26431799      25.45        110679 134339 111896 152478 182579 183790
+2           India                10777284      10.38         11039   8635  11436  11666  15244  16375
+3          Brazil                 9283418       8.94         54096  24591  27756  63520  64385  20006
+4  United Kingdom                 3852623       3.71         16840  18607  21088  25308  38905  58784
+5          Russia                 3842145       3.70         16406  17391  18083  17464  20806  23015
+6          France                 3213360       3.09         22659   4347  19080  26916  26497   4022
+7           Spain                 2851869       2.75         29064  79686      0  40285  41576  30579
+8           Italy                 2570608       2.47          9651   7925  11249  15191  13568  10798
+9          Turkey                 2492977       2.40          7795   7719   6562   7489   6435  13695
+10        Germany                 2240017       2.16          7690   6668    748  15636  29003  12320
+11       Colombia                 2114597       2.04         10091   9622   8078  13953  17908  10311
+12      Argentina                 1943548       1.87          9695   6614   4975  10843  12112   8222
+13         Mexico                 1874092       1.80          4384   5448   7030  17944  20548   6464
+14         Poland                 1520215       1.46          4326   2504   4711   6790   6943   4385
+15   South Africa                 1458958       1.40          2649   2548   4525   7070  12710  12601
 -------------------------------------------------------------------------------- 
[…]
