+++ NAMESPACE
@@ -34,7 +34,6 @@ export(mtrends)
 export(nullify.data)
 export(plt.SIR.model)
 export(preProcessingData)
-export(report.Tor)
 export(report.summary)
 export(rollingRate)
 export(single.trend)
