### 2021-02-03 7ee5214
pandemics loading fn binding to namespace
