+++ NAMESPACE
@@ -32,6 +32,7 @@ export(live.map)
 export(mrollingRates)
 export(mtrends)
 export(nullify.data)
+export(pandemics.data)
 export(plt.SIR.model)
 export(preProcessingData)
 export(report.summary)
