### 2021-02-03 03d2a59
daily update of covid19 data/plots/figures: covid19-SummaryReport.txt ... Wed Feb  3 13:03:47 EST 2021-gargantua
