+++ man/figures/covid19-SummaryReport.txt
@@ -1,16 +1,16 @@
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ 
 -------------------------------------------------------------------------------- 
 ################################################################################ 
-  ##### TS-CONFIRMED Cases  -- Data dated:  2021-02-02  ::  2021-02-03 01:02:36 
+  ##### TS-CONFIRMED Cases  -- Data dated:  2021-02-02  ::  2021-02-03 13:02:39 
 ################################################################################ 
   Number of Countries/Regions reported:  192 
   Number of Cities/Provinces reported:  85 
   Unique number of distinct geographical locations combined: 273 
 -------------------------------------------------------------------------------- 
-  Worldwide ts-confirmed  Totals: 103866068 
+  Worldwide ts-confirmed  Totals: 103869826 
 -------------------------------------------------------------------------------- 
    Country.Region Province.State   Totals GlobalPerc LastDayChange    t-2    t-3    t-7   t-14   t-30
-1              US                26431799      25.45        110679 134339 111896 152478 182579 183790
+1              US                26435557      25.45        114437 134339 111896 152478 182579 183790
 2           India                10777284      10.38         11039   8635  11436  11666  15244  16375
 3          Brazil                 9283418       8.94         54096  24591  27756  63520  64385  20006
 4  United Kingdom                 3852623       3.71         16840  18607  21088  25308  38905  58784
@@ -33,16 +33,16 @@
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ 
 -------------------------------------------------------------------------------- 
 ################################################################################ 
-  ##### TS-DEATHS Cases  -- Data dated:  2021-02-02  ::  2021-02-03 01:02:36 
+  ##### TS-DEATHS Cases  -- Data dated:  2021-02-02  ::  2021-02-03 13:02:39 
 ################################################################################ 
   Number of Countries/Regions reported:  192 
   Number of Cities/Provinces reported:  85 
   Unique number of distinct geographical locations combined: 273 
 -------------------------------------------------------------------------------- 
-  Worldwide ts-deaths  Totals: 2252923 
+  Worldwide ts-deaths  Totals: 2253064 
 -------------------------------------------------------------------------------- 
    Country.Region Province.State Totals Perc LastDayChange  t-2  t-3  t-7 t-14 t-30
-1              US                446744 1.69          3389 2031 1794 3943 4377 1995
+1              US                446885 1.69          3530 2031 1794 3943 4377 1995
 2          Brazil                226309 2.44          1210  595  559 1283 1340  543
 3          Mexico                159533 8.51           433  564  462 1623 1539  544
 4           India                154596 1.43           110   94  118  123  151  201
@@ -63,7 +63,7 @@
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ 
 -------------------------------------------------------------------------------- 
 ################################################################################ 
-  ##### TS-RECOVERED Cases  -- Data dated:  2021-02-02  ::  2021-02-03 01:02:37 
+  ##### TS-RECOVERED Cases  -- Data dated:  2021-02-02  ::  2021-02-03 13:02:40 
 ################################################################################ 
   Number of Countries/Regions reported:  192 
   Number of Cities/Provinces reported:  69 
@@ -92,7 +92,7 @@
[…]
