### 2021-02-04 7e1669f
daily update of covid19 data/plots/figures: covid19-SummaryReport.txt ... Thu Feb  4 01:03:38 EST 2021-gargantua
