+++ man/figures/covid19-SummaryReport.txt
@@ -1,202 +1,202 @@
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ 
 -------------------------------------------------------------------------------- 
 ################################################################################ 
-  ##### TS-CONFIRMED Cases  -- Data dated:  2021-02-02  ::  2021-02-03 13:02:39 
+  ##### TS-CONFIRMED Cases  -- Data dated:  2021-02-03  ::  2021-02-04 01:02:37 
 ################################################################################ 
   Number of Countries/Regions reported:  192 
   Number of Cities/Provinces reported:  85 
   Unique number of distinct geographical locations combined: 273 
 -------------------------------------------------------------------------------- 
-  Worldwide ts-confirmed  Totals: 103869826 
+  Worldwide ts-confirmed  Totals: 104389634 
 -------------------------------------------------------------------------------- 
    Country.Region Province.State   Totals GlobalPerc LastDayChange    t-2    t-3    t-7   t-14   t-30
-1              US                26435557      25.45        114437 134339 111896 152478 182579 183790
-2           India                10777284      10.38         11039   8635  11436  11666  15244  16375
-3          Brazil                 9283418       8.94         54096  24591  27756  63520  64385  20006
-4  United Kingdom                 3852623       3.71         16840  18607  21088  25308  38905  58784
-5          Russia                 3842145       3.70         16406  17391  18083  17464  20806  23015
-6          France                 3213360       3.09         22659   4347  19080  26916  26497   4022
-7           Spain                 2851869       2.75         29064  79686      0  40285  41576  30579
-8           Italy                 2570608       2.47          9651   7925  11249  15191  13568  10798
-9          Turkey                 2492977       2.40          7795   7719   6562   7489   6435  13695
-10        Germany                 2240017       2.16          7690   6668    748  15636  29003  12320
-11       Colombia                 2114597       2.04         10091   9622   8078  13953  17908  10311
-12      Argentina                 1943548       1.87          9695   6614   4975  10843  12112   8222
-13         Mexico                 1874092       1.80          4384   5448   7030  17944  20548   6464
-14         Poland                 1520215       1.46          4326   2504   4711   6790   6943   4385
-15   South Africa                 1458958       1.40          2649   2548   4525   7070  12710  12601
+1              US                26554792      25.44        119235 114437 134339 168620 193055 233526
+2           India                10790183      10.34         12899  11039   8635  18855  14545  18088
+3          Brazil                 9339420       8.95         56002  54096  24591  61811  59119  56648
+4  United Kingdom                 3871825       3.71         19202  16840  18607  28680  37892  60916
+5          Russia                 3858367       3.70         16222  16406  17391  18856  21544  23955
+6          France                 3239256       3.10         25896  22659   4347      0  22368  20280
+7           Spain                 2883465       2.76         31596  29064  79686  34899  44357  23700
+8           Italy                 2583790       2.48         13182   9651   7925  14360  14055  15375
+9          Turkey                 2501079       2.40          8102   7795   7719   7279   6289  14494
+10        Germany                 2252504       2.16         12487   7690   6668  14883   8277  18349
+11       Colombia                 2125622       2.04         11025  10091   9622  12270  15366  16835
+12      Argentina                 1952744       1.87          9196   9695   6614   9471  11396  13790
+13         Mexico                 1886245       1.81         12153   4384   5448  18670  22339  11271
+14         Poland                 1527016       1.46          6801   4326   2504   7153   7008   7596
+15   South Africa                 1463016       1.40          4058   2649   2548   7150  11381  14410
 -------------------------------------------------------------------------------- 
[…]
