+++ tests/testthat/test_examples.R
@@ -29,10 +29,12 @@ test_that("Data Retrieval", {
         expect_type(covid19.Toronto.data(origin="city",data.fmt="original"), "list")
 
         # TS Toronto data
-	# failing on WINDOWS i386
-	#tor.data <- covid19.Toronto.data()
-	#expect_s3_class(tor.data,"data.frame")
-        tor.data <- covid19.Toronto.data(local.data=TRUE)
+	# failing on WINDOWS ##i386
+	tor.data <- covid19.Toronto.data()
+	expect_s3_class(tor.data,"data.frame")
+
+	## failing on WINDOWS -- R 3.6.3
+	tor.data <- covid19.Toronto.data(local.data=TRUE)
         expect_s3_class(tor.data,"data.frame")
 
 	# US data
@@ -47,9 +49,11 @@ test_that("Data Retrieval", {
 
 test_that("Toronto Data", {
 	# TS Toronto data
-	# failing on WINDOWS i386
-	#tor.data <- covid19.Toronto.data()
-	#expect_s3_class(tor.data,"data.frame")
+	# failing on WINDOWS ##i386
+	tor.data <- covid19.Toronto.data()
+	expect_s3_class(tor.data,"data.frame")
+
+	# failing on WINDOWS -- R 3.6.3
 	tor.data <- covid19.Toronto.data(local.data=TRUE)
 	expect_s3_class(tor.data,"data.frame")
 
