### 2021-02-04 777c6cc
daily update of covid19 data/plots/figures: covid19-SummaryReport.txt ... Thu Feb  4 13:03:41 EST 2021-gargantua
