+++ man/figures/covid19-SummaryReport.txt
@@ -1,16 +1,16 @@
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ 
 -------------------------------------------------------------------------------- 
 ################################################################################ 
-  ##### TS-CONFIRMED Cases  -- Data dated:  2021-02-03  ::  2021-02-04 01:02:37 
+  ##### TS-CONFIRMED Cases  -- Data dated:  2021-02-03  ::  2021-02-04 13:02:38 
 ################################################################################ 
   Number of Countries/Regions reported:  192 
   Number of Cities/Provinces reported:  85 
   Unique number of distinct geographical locations combined: 273 
 -------------------------------------------------------------------------------- 
-  Worldwide ts-confirmed  Totals: 104389634 
+  Worldwide ts-confirmed  Totals: 104391868 
 -------------------------------------------------------------------------------- 
    Country.Region Province.State   Totals GlobalPerc LastDayChange    t-2    t-3    t-7   t-14   t-30
-1              US                26554792      25.44        119235 114437 134339 168620 193055 233526
+1              US                26557026      25.44        121469 114437 134339 168620 193055 233526
 2           India                10790183      10.34         12899  11039   8635  18855  14545  18088
 3          Brazil                 9339420       8.95         56002  54096  24591  61811  59119  56648
 4  United Kingdom                 3871825       3.71         19202  16840  18607  28680  37892  60916
@@ -33,16 +33,16 @@
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ 
 -------------------------------------------------------------------------------- 
 ################################################################################ 
-  ##### TS-DEATHS Cases  -- Data dated:  2021-02-03  ::  2021-02-04 01:02:38 
+  ##### TS-DEATHS Cases  -- Data dated:  2021-02-03  ::  2021-02-04 13:02:39 
 ################################################################################ 
   Number of Countries/Regions reported:  192 
   Number of Cities/Provinces reported:  85 
   Unique number of distinct geographical locations combined: 273 
 -------------------------------------------------------------------------------- 
-  Worldwide ts-deaths  Totals: 2268681 
+  Worldwide ts-deaths  Totals: 2268797 
 -------------------------------------------------------------------------------- 
    Country.Region Province.State Totals Perc LastDayChange  t-2  t-3  t-7 t-14 t-30
-1              US                450681 1.70          3796 3530 2031 4001 4203 3750
+1              US                450797 1.70          3912 3530 2031 4001 4203 3750
 2          Brazil                227563 2.44          1254 1210  595 1386 1316 1171
 3          Mexico                161240 8.55          1707  433  564 1506 1803 1065
 4           India                154703 1.43           107  110   94  163  163  264
@@ -63,7 +63,7 @@
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ 
 -------------------------------------------------------------------------------- 
 ################################################################################ 
-  ##### TS-RECOVERED Cases  -- Data dated:  2021-02-03  ::  2021-02-04 01:02:39 
+  ##### TS-RECOVERED Cases  -- Data dated:  2021-02-03  ::  2021-02-04 13:02:39 
 ################################################################################ 
   Number of Countries/Regions reported:  192 
[…]
