+++ man/figures/covid19-SummaryReport.txt
@@ -1,30 +1,30 @@
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ 
 -------------------------------------------------------------------------------- 
 ################################################################################ 
-  ##### TS-CONFIRMED Cases  -- Data dated:  2021-02-03  ::  2021-02-04 13:02:38 
+  ##### TS-CONFIRMED Cases  -- Data dated:  2021-02-04  ::  2021-02-05 01:02:37 
 ################################################################################ 
   Number of Countries/Regions reported:  192 
   Number of Cities/Provinces reported:  85 
   Unique number of distinct geographical locations combined: 273 
 -------------------------------------------------------------------------------- 
-  Worldwide ts-confirmed  Totals: 104391868 
+  Worldwide ts-confirmed  Totals: 104854731 
 -------------------------------------------------------------------------------- 
    Country.Region Province.State   Totals GlobalPerc LastDayChange    t-2    t-3    t-7   t-14   t-30
-1              US                26557026      25.44        121469 114437 134339 168620 193055 233526
-2           India                10790183      10.34         12899  11039   8635  18855  14545  18088
-3          Brazil                 9339420       8.95         56002  54096  24591  61811  59119  56648
-4  United Kingdom                 3871825       3.71         19202  16840  18607  28680  37892  60916
-5          Russia                 3858367       3.70         16222  16406  17391  18856  21544  23955
-6          France                 3239256       3.10         25896  22659   4347      0  22368  20280
-7           Spain                 2883465       2.76         31596  29064  79686  34899  44357  23700
-8           Italy                 2583790       2.48         13182   9651   7925  14360  14055  15375
-9          Turkey                 2501079       2.40          8102   7795   7719   7279   6289  14494
-10        Germany                 2252504       2.16         12487   7690   6668  14883   8277  18349
-11       Colombia                 2125622       2.04         11025  10091   9622  12270  15366  16835
-12      Argentina                 1952744       1.87          9196   9695   6614   9471  11396  13790
-13         Mexico                 1886245       1.81         12153   4384   5448  18670  22339  11271
-14         Poland                 1527016       1.46          6801   4326   2504   7153   7008   7596
-15   South Africa                 1463016       1.40          4058   2649   2548   7150  11381  14410
+1              US                26676957      25.44        119931 121469 114437 166113 190032 253917
+2           India                10802591      10.30         12408  12899  11039  13082  14256  20346
+3          Brazil                 9396293       8.96         56873  56002  54096  59826  56552  63430
+4  United Kingdom                 3892459       3.71         20634  19202  16840  29079  40261  62322
+5          Russia                 3874830       3.70         16463  16222  16406  18966  21182  23902
+6          France                 3238798       3.09          -458  25896  22659  46043  22995  25020
+7           Spain                 2913425       2.78         29960  31596  29064  38118  42885      0
+8           Italy                 2597446       2.48         13656  13182   9651  13563  13633  20326
+9          Turkey                 2508988       2.39          7909   8102   7795   6912   5967  13830
+10        Germany                 2265536       2.16         13032  12487   7690  12831  16366  26663
+11       Colombia                 2135412       2.04          9790  11025  10091  10058  15073  16805
+12      Argentina                 1961635       1.87          8891   9196   9695   9838  10753  13441
+13         Mexico                 1886245       1.80             0  12153   4384  16374  21007  13345
+14         Poland                 1533511       1.46          6495   6801   4326   6145   6693  14220
+15   South Africa                 1466767       1.40          3751   4058   2649   6141  11761  21832
 -------------------------------------------------------------------------------- 
[…]
