### 2021-02-05 8599aec
update travis.ci to skip builds for daily updates
