### 2021-02-05 14672e8
daily update of covid19 data/plots/figures: covid19-SummaryReport.txt ... Fri Feb  5 13:03:47 EST 2021-gargantua
