+++ man/figures/covid19-SummaryReport.txt
@@ -1,16 +1,16 @@
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ 
 -------------------------------------------------------------------------------- 
 ################################################################################ 
-  ##### TS-CONFIRMED Cases  -- Data dated:  2021-02-04  ::  2021-02-05 01:02:37 
+  ##### TS-CONFIRMED Cases  -- Data dated:  2021-02-04  ::  2021-02-05 13:02:41 
 ################################################################################ 
   Number of Countries/Regions reported:  192 
   Number of Cities/Provinces reported:  85 
   Unique number of distinct geographical locations combined: 273 
 -------------------------------------------------------------------------------- 
-  Worldwide ts-confirmed  Totals: 104854731 
+  Worldwide ts-confirmed  Totals: 104857988 
 -------------------------------------------------------------------------------- 
    Country.Region Province.State   Totals GlobalPerc LastDayChange    t-2    t-3    t-7   t-14   t-30
-1              US                26676957      25.44        119931 121469 114437 166113 190032 253917
+1              US                26680214      25.44        123188 121469 114437 166113 190032 253917
 2           India                10802591      10.30         12408  12899  11039  13082  14256  20346
 3          Brazil                 9396293       8.96         56873  56002  54096  59826  56552  63430
 4  United Kingdom                 3892459       3.71         20634  19202  16840  29079  40261  62322
@@ -33,16 +33,16 @@
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ 
 -------------------------------------------------------------------------------- 
 ################################################################################ 
-  ##### TS-DEATHS Cases  -- Data dated:  2021-02-04  ::  2021-02-05 01:02:37 
+  ##### TS-DEATHS Cases  -- Data dated:  2021-02-04  ::  2021-02-05 13:02:41 
 ################################################################################ 
   Number of Countries/Regions reported:  192 
   Number of Cities/Provinces reported:  85 
   Unique number of distinct geographical locations combined: 273 
 -------------------------------------------------------------------------------- 
-  Worldwide ts-deaths  Totals: 2282535 
+  Worldwide ts-deaths  Totals: 2282678 
 -------------------------------------------------------------------------------- 
    Country.Region Province.State Totals Perc LastDayChange  t-2  t-3  t-7 t-14 t-30
-1              US                455738 1.71          4941 3912 3530 3603 3760 3890
+1              US                455881 1.71          5084 3912 3530 3603 3760 3890
 2          Brazil                228795 2.43          1232 1254 1210 1119 1096 1242
 3          Mexico                161240 8.55             0 1707  433 1434 1440 1165
 4           India                154823 1.43           120  107  110  137  152  222
@@ -63,7 +63,7 @@
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ 
 -------------------------------------------------------------------------------- 
 ################################################################################ 
-  ##### TS-RECOVERED Cases  -- Data dated:  2021-02-04  ::  2021-02-05 01:02:38 
+  ##### TS-RECOVERED Cases  -- Data dated:  2021-02-04  ::  2021-02-05 13:02:42 
 ################################################################################ 
   Number of Countries/Regions reported:  192 
[…]
