### 2021-02-06 df7172e
daily update of covid19 data/plots/figures: covid19-SummaryReport.txt ... Sat Feb  6 01:03:34 EST 2021-gargantua
