+++ man/figures/covid19-SummaryReport.txt
@@ -1,30 +1,30 @@
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ 
 -------------------------------------------------------------------------------- 
 ################################################################################ 
-  ##### TS-CONFIRMED Cases  -- Data dated:  2021-02-04  ::  2021-02-05 13:02:41 
+  ##### TS-CONFIRMED Cases  -- Data dated:  2021-02-05  ::  2021-02-06 01:02:38 
 ################################################################################ 
   Number of Countries/Regions reported:  192 
   Number of Cities/Provinces reported:  85 
   Unique number of distinct geographical locations combined: 273 
 -------------------------------------------------------------------------------- 
-  Worldwide ts-confirmed  Totals: 104857988 
+  Worldwide ts-confirmed  Totals: 105387133 
 -------------------------------------------------------------------------------- 
    Country.Region Province.State   Totals GlobalPerc LastDayChange    t-2    t-3    t-7   t-14   t-30
-1              US                26680214      25.44        123188 121469 114437 166113 190032 253917
-2           India                10802591      10.30         12408  12899  11039  13082  14256  20346
-3          Brazil                 9396293       8.96         56873  56002  54096  59826  56552  63430
-4  United Kingdom                 3892459       3.71         20634  19202  16840  29079  40261  62322
-5          Russia                 3874830       3.70         16463  16222  16406  18966  21182  23902
-6          France                 3238798       3.09          -458  25896  22659  46043  22995  25020
-7           Spain                 2913425       2.78         29960  31596  29064  38118  42885      0
-8           Italy                 2597446       2.48         13656  13182   9651  13563  13633  20326
-9          Turkey                 2508988       2.39          7909   8102   7795   6912   5967  13830
-10        Germany                 2265536       2.16         13032  12487   7690  12831  16366  26663
-11       Colombia                 2135412       2.04          9790  11025  10091  10058  15073  16805
-12      Argentina                 1961635       1.87          8891   9196   9695   9838  10753  13441
-13         Mexico                 1886245       1.80             0  12153   4384  16374  21007  13345
-14         Poland                 1533511       1.46          6495   6801   4326   6145   6693  14220
-15   South Africa                 1466767       1.40          3751   4058   2649   6141  11761  21832
+1              US                26808328      25.44        128114 123188 121469 142091 170138 276738
+2           India                10814304      10.26         11713  12408  12899  13044  14849  18139
+3          Brazil                 9447165       8.96         50872  56873  56002  58462  62334  87843
+4  United Kingdom                 3911573       3.71         19114  20634  19202  23275  33552  52618
+5          Russia                 3891274       3.69         16444  16463  16222  18751  20585  23218
+6          France                 3284085       3.12         45287   -458  25896  23857  23410      0
+7           Spain                 2941990       2.79         28565  29960  31596      0      0  42360
+8           Italy                 2611659       2.48         14213  13656  13182  12713  13331  18416
+9          Turkey                 2516889       2.39          7901   7909   8102   6871   5856  12171
+10        Germany                 2276371       2.16         10835  13032  12487  17518  12430  45333
+11       Colombia                 2142660       2.03          7248   9790  11025   9173  15551  17576
+12      Argentina                 1970009       1.87          8374   8891   9196   6902   8362  13835
+13         Mexico                 1912871       1.82         26626      0  12153  15337  20057  13734
+14         Poland                 1539564       1.46          6053   6495   6801   5864   6431  12119
+15   South Africa                 1470516       1.40          3749   3751   4058   5297  12271  20999
 -------------------------------------------------------------------------------- 
[…]
