### 2021-02-06 00c992f
daily update of covid19 data/plots/figures: covid19-SummaryReport.txt ... Sat Feb  6 13:03:36 EST 2021-gargantua
