+++ man/figures/covid19-SummaryReport.txt
@@ -1,16 +1,16 @@
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ 
 -------------------------------------------------------------------------------- 
 ################################################################################ 
-  ##### TS-CONFIRMED Cases  -- Data dated:  2021-02-05  ::  2021-02-06 01:02:38 
+  ##### TS-CONFIRMED Cases  -- Data dated:  2021-02-05  ::  2021-02-06 13:02:30 
 ################################################################################ 
   Number of Countries/Regions reported:  192 
   Number of Cities/Provinces reported:  85 
   Unique number of distinct geographical locations combined: 273 
 -------------------------------------------------------------------------------- 
-  Worldwide ts-confirmed  Totals: 105387133 
+  Worldwide ts-confirmed  Totals: 105392577 
 -------------------------------------------------------------------------------- 
    Country.Region Province.State   Totals GlobalPerc LastDayChange    t-2    t-3    t-7   t-14   t-30
-1              US                26808328      25.44        128114 123188 121469 142091 170138 276738
+1              US                26813772      25.44        133558 123188 121469 142091 170138 276738
 2           India                10814304      10.26         11713  12408  12899  13044  14849  18139
 3          Brazil                 9447165       8.96         50872  56873  56002  58462  62334  87843
 4  United Kingdom                 3911573       3.71         19114  20634  19202  23275  33552  52618
@@ -22,27 +22,27 @@
 10        Germany                 2276371       2.16         10835  13032  12487  17518  12430  45333
 11       Colombia                 2142660       2.03          7248   9790  11025   9173  15551  17576
 12      Argentina                 1970009       1.87          8374   8891   9196   6902   8362  13835
-13         Mexico                 1912871       1.82         26626      0  12153  15337  20057  13734
+13         Mexico                 1912871       1.81         26626      0  12153  15337  20057  13734
 14         Poland                 1539564       1.46          6053   6495   6801   5864   6431  12119
 15   South Africa                 1470516       1.40          3749   3751   4058   5297  12271  20999
 -------------------------------------------------------------------------------- 
   Global Perc. Average:  0.37 (sd: 1.81) 
-  Global Perc. Average in top  15 :  4.91 (sd: 6.26) 
+  Global Perc. Average in top  15 :  4.9 (sd: 6.26) 
 -------------------------------------------------------------------------------- 
 ================================================================================ 
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ 
 -------------------------------------------------------------------------------- 
 ################################################################################ 
-  ##### TS-DEATHS Cases  -- Data dated:  2021-02-05  ::  2021-02-06 01:02:39 
+  ##### TS-DEATHS Cases  -- Data dated:  2021-02-05  ::  2021-02-06 13:02:31 
 ################################################################################ 
   Number of Countries/Regions reported:  192 
   Number of Cities/Provinces reported:  85 
   Unique number of distinct geographical locations combined: 273 
 -------------------------------------------------------------------------------- 
-  Worldwide ts-deaths  Totals: 2299083 
+  Worldwide ts-deaths  Totals: 2299235 
 -------------------------------------------------------------------------------- 
    Country.Region Province.State Totals Perc LastDayChange  t-2  t-3  t-7 t-14 t-30
-1              US                459403 1.71          3522 5084 3912 2731 3329 3950
+1              US                459555 1.71          3674 5084 3912 2731 3329 3950
[…]
