### 2021-02-07 eb1e521
daily update of covid19 data/plots/figures: covid19-SummaryReport.txt ... Sun Feb  7 01:03:31 EST 2021-gargantua  [skip ci]
