+++ man/figures/covid19-SummaryReport.txt
@@ -1,30 +1,30 @@
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ 
 -------------------------------------------------------------------------------- 
 ################################################################################ 
-  ##### TS-CONFIRMED Cases  -- Data dated:  2021-02-05  ::  2021-02-06 13:02:30 
+  ##### TS-CONFIRMED Cases  -- Data dated:  2021-02-06  ::  2021-02-07 01:02:28 
 ################################################################################ 
   Number of Countries/Regions reported:  192 
   Number of Cities/Provinces reported:  85 
   Unique number of distinct geographical locations combined: 273 
 -------------------------------------------------------------------------------- 
-  Worldwide ts-confirmed  Totals: 105392577 
+  Worldwide ts-confirmed  Totals: 105763135 
 -------------------------------------------------------------------------------- 
    Country.Region Province.State   Totals GlobalPerc LastDayChange    t-2    t-3    t-7   t-14   t-30
-1              US                26813772      25.44        133558 123188 121469 142091 170138 276738
-2           India                10814304      10.26         11713  12408  12899  13044  14849  18139
-3          Brazil                 9447165       8.96         50872  56873  56002  58462  62334  87843
-4  United Kingdom                 3911573       3.71         19114  20634  19202  23275  33552  52618
-5          Russia                 3891274       3.69         16444  16463  16222  18751  20585  23218
-6          France                 3284085       3.12         45287   -458  25896  23857  23410      0
-7           Spain                 2941990       2.79         28565  29960  31596      0      0  42360
-8           Italy                 2611659       2.48         14213  13656  13182  12713  13331  18416
-9          Turkey                 2516889       2.39          7901   7909   8102   6871   5856  12171
-10        Germany                 2276371       2.16         10835  13032  12487  17518  12430  45333
-11       Colombia                 2142660       2.03          7248   9790  11025   9173  15551  17576
-12      Argentina                 1970009       1.87          8374   8891   9196   6902   8362  13835
-13         Mexico                 1912871       1.81         26626      0  12153  15337  20057  13734
-14         Poland                 1539564       1.46          6053   6495   6801   5864   6431  12119
-15   South Africa                 1470516       1.40          3749   3751   4058   5297  12271  20999
+1              US                26916192      25.45        102420 133558 123188 111896 131075 292044
+2           India                10826363      10.24         12059  11713  12408  11436  13203      0
+3          Brazil                 9447165       8.93             0  50872  56873  27756  28323  52035
+4  United Kingdom                 3929835       3.72         18262  19114  20634  21088  30004  68053
+5          Russia                 3907653       3.69         16379  16444  16463  18083  20800  23330
+6          France                 3303273       3.12         19188  45287   -458  19080  18038  41134
+7           Spain                 2941990       2.78             0  28565  29960      0      0  25456
+8           Italy                 2625098       2.48         13439  14213  13656  11249  11628  17529
+9          Turkey                 2524786       2.39          7897   7901   7909   6562   5277  11479
+10        Germany                 2285003       2.16          8632  10835  13032    748  10078  19077
+11       Colombia                 2151207       2.03          8547   7248   9790   8078  12516  18221
+12      Argentina                 1976689       1.87          6680   8374   8891   4975   5031  13346
+13         Mexico                 1926080       1.82         13209  26626      0   7030  10872  14362
+14         Poland                 1545530       1.46          5966   6053   6495   4711   4566   8763
+15   South Africa                 1473700       1.39          3184   3749   3751   4525   8147  21980
 -------------------------------------------------------------------------------- 
[…]
