### 2021-02-07 0913b55
daily update of covid19 data/plots/figures: covid19-SummaryReport.txt ... Sun Feb  7 13:03:31 EST 2021-gargantua  [skip ci]
