+++ man/figures/covid19-SummaryReport.txt
@@ -1,16 +1,16 @@
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ 
 -------------------------------------------------------------------------------- 
 ################################################################################ 
-  ##### TS-CONFIRMED Cases  -- Data dated:  2021-02-06  ::  2021-02-07 01:02:28 
+  ##### TS-CONFIRMED Cases  -- Data dated:  2021-02-06  ::  2021-02-07 13:02:28 
 ################################################################################ 
   Number of Countries/Regions reported:  192 
   Number of Cities/Provinces reported:  85 
   Unique number of distinct geographical locations combined: 273 
 -------------------------------------------------------------------------------- 
-  Worldwide ts-confirmed  Totals: 105763135 
+  Worldwide ts-confirmed  Totals: 105764730 
 -------------------------------------------------------------------------------- 
    Country.Region Province.State   Totals GlobalPerc LastDayChange    t-2    t-3    t-7   t-14   t-30
-1              US                26916192      25.45        102420 133558 123188 111896 131075 292044
+1              US                26917787      25.45        104015 133558 123188 111896 131075 292044
 2           India                10826363      10.24         12059  11713  12408  11436  13203      0
 3          Brazil                 9447165       8.93             0  50872  56873  27756  28323  52035
 4  United Kingdom                 3929835       3.72         18262  19114  20634  21088  30004  68053
@@ -33,16 +33,16 @@
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ 
 -------------------------------------------------------------------------------- 
 ################################################################################ 
-  ##### TS-DEATHS Cases  -- Data dated:  2021-02-06  ::  2021-02-07 01:02:29 
+  ##### TS-DEATHS Cases  -- Data dated:  2021-02-06  ::  2021-02-07 13:02:29 
 ################################################################################ 
   Number of Countries/Regions reported:  192 
   Number of Cities/Provinces reported:  85 
   Unique number of distinct geographical locations combined: 273 
 -------------------------------------------------------------------------------- 
-  Worldwide ts-deaths  Totals: 2309350 
+  Worldwide ts-deaths  Totals: 2309346 
 -------------------------------------------------------------------------------- 
    Country.Region Province.State Totals Perc LastDayChange  t-2  t-3  t-7 t-14 t-30
-1              US                462173 1.72          2618 3674 5084 1794 1775 4037
+1              US                462169 1.72          2614 3674 5084 1794 1775 4037
 2          Brazil                230034 2.43             0 1239 1232  559  592  962
 3          Mexico                165786 8.61          1496 3050    0  462  530 1038
 4           India                154996 1.43            78   95  120  118  131    0
@@ -63,7 +63,7 @@
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ 
 -------------------------------------------------------------------------------- 
 ################################################################################ 
-  ##### TS-RECOVERED Cases  -- Data dated:  2021-02-06  ::  2021-02-07 01:02:30 
+  ##### TS-RECOVERED Cases  -- Data dated:  2021-02-06  ::  2021-02-07 13:02:29 
 ################################################################################ 
   Number of Countries/Regions reported:  192 
   Number of Cities/Provinces reported:  69 
@@ -92,11 +92,11 @@
[…]
