### 2021-02-08 d0d75a0
daily update of covid19 data/plots/figures: covid19-SummaryReport.txt ... Mon Feb  8 01:03:30 EST 2021-gargantua  [skip ci]
