+++ man/figures/covid19-SummaryReport.txt
@@ -1,30 +1,30 @@
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ 
 -------------------------------------------------------------------------------- 
 ################################################################################ 
-  ##### TS-CONFIRMED Cases  -- Data dated:  2021-02-06  ::  2021-02-07 13:02:28 
+  ##### TS-CONFIRMED Cases  -- Data dated:  2021-02-07  ::  2021-02-08 01:02:27 
 ################################################################################ 
   Number of Countries/Regions reported:  192 
   Number of Cities/Provinces reported:  85 
   Unique number of distinct geographical locations combined: 273 
 -------------------------------------------------------------------------------- 
-  Worldwide ts-confirmed  Totals: 105764730 
+  Worldwide ts-confirmed  Totals: 106159883 
 -------------------------------------------------------------------------------- 
    Country.Region Province.State   Totals GlobalPerc LastDayChange    t-2    t-3    t-7   t-14   t-30
-1              US                26917787      25.45        104015 133558 123188 111896 131075 292044
-2           India                10826363      10.24         12059  11713  12408  11436  13203      0
-3          Brazil                 9447165       8.93             0  50872  56873  27756  28323  52035
-4  United Kingdom                 3929835       3.72         18262  19114  20634  21088  30004  68053
-5          Russia                 3907653       3.69         16379  16444  16463  18083  20800  23330
-6          France                 3303273       3.12         19188  45287   -458  19080  18038  41134
-7           Spain                 2941990       2.78             0  28565  29960      0      0  25456
-8           Italy                 2625098       2.48         13439  14213  13656  11249  11628  17529
-9          Turkey                 2524786       2.39          7897   7901   7909   6562   5277  11479
-10        Germany                 2285003       2.16          8632  10835  13032    748  10078  19077
-11       Colombia                 2151207       2.03          8547   7248   9790   8078  12516  18221
-12      Argentina                 1976689       1.87          6680   8374   8891   4975   5031  13346
-13         Mexico                 1926080       1.82         13209  26626      0   7030  10872  14362
-14         Poland                 1545530       1.46          5966   6053   6495   4711   4566   8763
-15   South Africa                 1473700       1.39          3184   3749   3751   4525   8147  21980
+1              US                27004715      25.44         86928 104015 133558 134339 151095 261974
+2           India                10838194      10.21         11831  12059  11713   8635   9102  36867
+3          Brazil                 9524640       8.97         77475      0  50872  24591  26816  62290
+4  United Kingdom                 3945680       3.72         15845  18262  19114  18607  22195  59937
+5          Russia                 3923461       3.70         15808  16379  16444  17391  18999  23012
+6          France                 3322988       3.13         19715  19188  45287   4347   4240  19771
+7           Spain                 2941990       2.77             0      0  28565  79686  93822      0
+8           Italy                 2636738       2.48         11640  13439  14213   7925   8559  19976
+9          Turkey                 2531456       2.38          6670   7897   7901   7719   5642   9537
+10        Germany                 2291673       2.16          6670   8632  10835   6668   6887  22824
+11       Colombia                 2157216       2.03          6009   8547   7248   9622  12261  15795
+12      Argentina                 1980347       1.87          3658   6680   8374   6614   7578  11057
+13         Mexico                 1932145       1.82          6065  13209  26626   5448   8521  16105
+14         Poland                 1550255       1.46          4725   5966   6053   2504   2674  10744
+15   South Africa                 1476135       1.39          2435   3184   3749   2548   4551  21606
 -------------------------------------------------------------------------------- 
[…]
