### 2021-02-08 488360c
daily update of covid19 data/plots/figures: covid19-SummaryReport.txt ... Mon Feb  8 13:03:26 EST 2021-gargantua  [skip ci]
