+++ man/figures/covid19-SummaryReport.txt
@@ -1,16 +1,16 @@
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ 
 -------------------------------------------------------------------------------- 
 ################################################################################ 
-  ##### TS-CONFIRMED Cases  -- Data dated:  2021-02-07  ::  2021-02-08 01:02:27 
+  ##### TS-CONFIRMED Cases  -- Data dated:  2021-02-07  ::  2021-02-08 13:02:26 
 ################################################################################ 
   Number of Countries/Regions reported:  192 
   Number of Cities/Provinces reported:  85 
   Unique number of distinct geographical locations combined: 273 
 -------------------------------------------------------------------------------- 
-  Worldwide ts-confirmed  Totals: 106159883 
+  Worldwide ts-confirmed  Totals: 106160999 
 -------------------------------------------------------------------------------- 
    Country.Region Province.State   Totals GlobalPerc LastDayChange    t-2    t-3    t-7   t-14   t-30
-1              US                27004715      25.44         86928 104015 133558 134339 151095 261974
+1              US                27005831      25.44         88044 104015 133558 134339 151095 261974
 2           India                10838194      10.21         11831  12059  11713   8635   9102  36867
 3          Brazil                 9524640       8.97         77475      0  50872  24591  26816  62290
 4  United Kingdom                 3945680       3.72         15845  18262  19114  18607  22195  59937
@@ -33,16 +33,16 @@
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ 
 -------------------------------------------------------------------------------- 
 ################################################################################ 
-  ##### TS-DEATHS Cases  -- Data dated:  2021-02-07  ::  2021-02-08 01:02:28 
+  ##### TS-DEATHS Cases  -- Data dated:  2021-02-07  ::  2021-02-08 13:02:27 
 ################################################################################ 
   Number of Countries/Regions reported:  192 
   Number of Cities/Provinces reported:  85 
   Unique number of distinct geographical locations combined: 273 
 -------------------------------------------------------------------------------- 
-  Worldwide ts-deaths  Totals: 2317162 
+  Worldwide ts-deaths  Totals: 2317170 
 -------------------------------------------------------------------------------- 
    Country.Region Province.State Totals Perc LastDayChange  t-2  t-3  t-7 t-14 t-30
-1              US                463437 1.72          1268 2614 3674 2031 1917 3251
+1              US                463445 1.72          1276 2614 3674 2031 1917 3251
 2          Brazil                231534 2.43          1500    0 1239  595  627 1171
 3          Mexico                166200 8.60           414 1496 3050  564  659 1135
 4           India                155080 1.43            84   78   95   94  117  429
@@ -63,7 +63,7 @@
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ 
 -------------------------------------------------------------------------------- 
 ################################################################################ 
-  ##### TS-RECOVERED Cases  -- Data dated:  2021-02-07  ::  2021-02-08 01:02:29 
+  ##### TS-RECOVERED Cases  -- Data dated:  2021-02-07  ::  2021-02-08 13:02:27 
 ################################################################################ 
[…]
