### 2021-02-08 #2995 @JDRomano2
Thanks very much @fboehm for the invitation to review this submission! Please see below for individual comments on each of the checklist items that I feel need attention:

## General checks
### Contribution and authorship: Has the submitting author (@mponce0) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?

The contributions made by the second author are not clear, as they are not a contributor to the GitHub source repository, and their contributions are not described elsewhere in the paper. I would strongly recommend including information about author contributions in an "About" heading within the GitHub README, as well as current contact information for at least the primary author.

## Documentation
### Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?

Documentation is provided primarily by means of examples. No API documentation is given.

If a reference manual with API documentation is available on CRAN, there should be a link to it from the GitHub README and/or the project website.

### Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

I cannot find any community guidelines in the paper, the code repository, or the project's website.

The process for reporting bugs and issues is especially important to include, and should be clearly defined in the README. Ideally this would make use of GitHub's issue tracking functionality.

## Software paper
### Summary: Has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided?

The summary is somewhat vague and nonspecific.

> "With the emergence of a new pandemic worldwide, a novel strategy to approach it has also emerged."

What does 'approach it' mean? The paper should clearly describe the types of data-related problems that the software is meant to solve.

### A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?

The authors do describe the motivation behind the project in the context of other open-source COVID-19 data resources. However, the "Statement of Need" section needs to be cleaned up. Much of the text in here should go either into the Summary (keeping in mind the recommended edits mentioned above) or into a separate section following the Statement of Need describing the software's functionality.

The GitHub README file contains a fairly good description of the needs this software is meant to address, and some of that text can probably be reused in the paper.

Other section headings should be used to denote where the "statement of need" ends and where other topics begin.

### Quality of writing: Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?

The paper is in need of stylistic and typographical revisions. A few examples (not comprehensive) include:

- "Coronavirus"; not "Corona virus"
- "COVID-19"; not "CoViD19"
- "GitHub"; not "GitHUB"
- Capitalized letters at the beginning of each item in the bulleted lists
- "the R language"; not "the R Language and Environment for Statistical Computing" (which is just the title of the publication, not the actual language's title).
- Why is the word "backup" in bold on the first page?

These are just a few examples that stand out.

### References: Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper citation syntax?

Some of the citations are incomplete. For example, the citation for "The Comprehensive R Archive Network" should include the journal name, author, publication date, volume, and issue number. These are easily found by clicking on the DOI link provided in the reference. Similar omissions are found in several of the other references.

I'm not sure that enough of the citations point to appropriate publications or data sources. For example, you list the Visual Capitalist infographic as one of the data sources used by your software, but the linked article doesn't seem to provide any way to access the data.
