### 2021-02-09 1375e1e
daily update of covid19 data/plots/figures: covid19-SummaryReport.txt ... Tue Feb  9 01:03:32 EST 2021-gargantua  [skip ci]
