+++ man/figures/covid19-SummaryReport.txt
@@ -1,30 +1,30 @@
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ 
 -------------------------------------------------------------------------------- 
 ################################################################################ 
-  ##### TS-CONFIRMED Cases  -- Data dated:  2021-02-07  ::  2021-02-08 13:02:26 
+  ##### TS-CONFIRMED Cases  -- Data dated:  2021-02-08  ::  2021-02-09 01:02:28 
 ################################################################################ 
   Number of Countries/Regions reported:  192 
   Number of Cities/Provinces reported:  85 
   Unique number of distinct geographical locations combined: 273 
 -------------------------------------------------------------------------------- 
-  Worldwide ts-confirmed  Totals: 106160999 
+  Worldwide ts-confirmed  Totals: 106474965 
 -------------------------------------------------------------------------------- 
-   Country.Region Province.State   Totals GlobalPerc LastDayChange    t-2    t-3    t-7   t-14   t-30
-1              US                27005831      25.44         88044 104015 133558 134339 151095 261974
-2           India                10838194      10.21         11831  12059  11713   8635   9102  36867
-3          Brazil                 9524640       8.97         77475      0  50872  24591  26816  62290
-4  United Kingdom                 3945680       3.72         15845  18262  19114  18607  22195  59937
-5          Russia                 3923461       3.70         15808  16379  16444  17391  18999  23012
-6          France                 3322988       3.13         19715  19188  45287   4347   4240  19771
-7           Spain                 2941990       2.77             0      0  28565  79686  93822      0
-8           Italy                 2636738       2.48         11640  13439  14213   7925   8559  19976
-9          Turkey                 2531456       2.38          6670   7897   7901   7719   5642   9537
-10        Germany                 2291673       2.16          6670   8632  10835   6668   6887  22824
-11       Colombia                 2157216       2.03          6009   8547   7248   9622  12261  15795
-12      Argentina                 1980347       1.87          3658   6680   8374   6614   7578  11057
-13         Mexico                 1932145       1.82          6065  13209  26626   5448   8521  16105
-14         Poland                 1550255       1.46          4725   5966   6053   2504   2674  10744
-15   South Africa                 1476135       1.39          2435   3184   3749   2548   4551  21606
+   Country.Region Province.State   Totals GlobalPerc LastDayChange   t-2    t-3    t-7   t-14   t-30
+1              US                27094014      25.45         86646 89581 104015 114437 146597 213281
+2           India                10847304      10.19          9110 11831  12059  11039  12689  16311
+3          Brazil                 9524640       8.95             0 77475      0  54096  61963  29792
+4  United Kingdom                 3959784       3.72         14104 15845  18262  16840  20088  54940
+5          Russia                 3939162       3.70         15701 15808  16379  16406  17982  22540
+6          France                 3327305       3.12          4317 19715  19188  22659  21860  15801
+7           Spain                 2989085       2.81         47095     0      0  29064  36435      0
+8           Italy                 2644707       2.48          7969 11640  13439   9651  10584  18625
+9          Turkey                 2539559       2.39          8103  6670   7897   7795   7103   9138
+10        Germany                 2296323       2.16          4650  6670   8632   7690   9387    948
+11       Colombia                 2161462       2.03          4246  6009   8547  10091  13606  15537
+12      Argentina                 1985501       1.86          5154  3658   6680   9695  10409   7808
+13         Mexico                 1936013       1.82          3868  6065  13209   4384  17165  10003
+14         Poland                 1552686       1.46          2431  4725   5966   4326   4603   9133
+15   South Africa                 1477511       1.39          1376  2435   3184   2649   6041  17421
[…]
