### 2021-02-09 f476929
Adding about and authors section
