### 2021-02-09 67bed00
stylistic and typographical revisions, capitalizations, etc...
