### 2021-02-09 88ba647
completing entry for CRAN
