+++ JOSS/references.bib
@@ -23,10 +23,15 @@
 
 %%%%%%%%%%%%%%%%
 
-@Manual{cran,
-  title = {{The Comprehensive R Archive Network}},
-  Doi = {10.1002/wics.1212},
-  url = {https://cran.r-project.org}
+@article{cran,
+	author = {Hornik, Kurt},
+	title = {The Comprehensive R Archive Network},
+	journal = {WIREs Computational Statistics },
+	volume = {4},
+	number = {4},
+	pages = {394-398},
+	doi = {10.1002/wics.1212},
+	year = {2012}
 }
 
 %%%%%%%%%%%%%%%%
