### 2021-02-09 #2995 @mponce0
Many thanks @JDRomano2 for the through review and useful comments!
The editions and updates addressing this reviewer's comments have been added to the manuscript on the "literature" branch of the  https://github.com/mponce0/covid19.analytics/  repository.

> ## General checks
> ### Contribution and authorship: Has the submitting author (@mponce0) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?
> The contributions made by the second author are not clear, as they are not a contributor to the GitHub source repository, and their contributions are not described elsewhere in the paper. I would strongly recommend including information about author contributions in an "About" heading within the GitHub README, as well as current contact information for at least the primary author.

Authors contributions were (and are) originally described in the "Description" file, which is the standard place for doing it in R packages.

Following the reviewer's recommendation we have also added an "About" heading to the README document stating more clearly the contributions of the authors, as well as the way for the community to contribute to the package.

> ## Documentation
> ### Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?
> Documentation is provided primarily by means of examples. No API documentation is given.
>
> If a reference manual with API documentation is available on CRAN, there should be a link to it from the GitHub README and/or the project website.

A link has been added to the README for API documentation, in this case, functions descriptions from CRAN immediately after the ``Overview of the Main Functions from the "covid19.analytics" Package`` section.

> ### Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support
> I cannot find any community guidelines in the paper, the code repository, or the project's website.
>
> The process for reporting bugs and issues is especially important to include, and should be clearly defined in the README. Ideally this would make use of GitHub's issue tracking functionality.

This was mentioned both in the paper as well as in the README file, in the first bullet point of "Statement of need":
``Users can also submit tickets for bugs, suggestions or comments using the "issues" tab`` --referring to the GitHub repository--.
We have also added a paragraph to the paper about how to seek for support using the ISSUES tracker and contribute via pull request.
As a side note please notice that both features have been in place and used, as you can verify in the repo (https://github.com/mponce0/covid19.analytics/issues) since quite some time ago.

> ## Software paper
> ### Summary: Has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided?
> The summary is somewhat vague and nonspecific.
>
> > "With the emergence of a new pandemic worldwide, a novel strategy to approach it has also emerged."
>
> What does 'approach it' mean? The paper should clearly describe the types of data-related problems that the software is meant to solve.

We have clarified the summary and added a few sentences to describe what the package can allow users to achieve in terms of the data availability and its challenges.

> ### A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?
> The authors do describe the motivation behind the project in the context of other open-source COVID-19 data resources. However, the "Statement of Need" section needs to be cleaned up. Much of the text in here should go either into the Summary (keeping in mind the recommended edits mentioned above) or into a separate section following the Statement of Need describing the software's functionality.
>
> The GitHub README file contains a fairly good description of the needs this software is meant to address, and some of that text can probably be reused in the paper.

We have restructured and edited the manuscript to incorporate the reviewer's comments.

> ### Quality of writing: Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?
> The paper is in need of stylistic and typographical revisions. A few examples (not comprehensive) include:
>
> * "Coronavirus"; not "Corona virus"
> * "COVID-19"; not "CoViD19"
> * "GitHub"; not "GitHUB"
> * Capitalized letters at the beginning of each item in the bulleted lists
> * "the R language"; not "the R Language and Environment for Statistical Computing" (which is just the title of the publication, not the actual language's title).
> * Why is the word "backup" in bold on the first page?
>
> These are just a few examples that stand out.

We thank the reviewer for noticing this, and we have revisit and edited the manuscript to fix all of these.

> ### References: Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper citation syntax?
> Some of the citations are incomplete. For example, the citation for "The Comprehensive R Archive Network" should include the journal name, author, publication date, volume, and issue number. These are easily found by clicking on the DOI link provided in the reference. Similar omissions are found in several of the other references.

We have reviewed the citations and completed the information for the CRAN publication. We couldn't identify any other reference missing any of the elements described above.

> I'm not sure that enough of the citations point to appropriate publications or data sources. For example, you list the Visual Capitalist infographic as one of the data sources used by your software, but the linked article doesn't seem to provide any way to access the data.

This is because the data from the infographics was gather manually, ie. no actual data repository is offered.
WRT the other data sources, we will be OK to follow any other recommendation in how to better reference/cite these data sources.
