+++ man/figures/covid19-SummaryReport.txt
@@ -1,202 +1,202 @@
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ 
 -------------------------------------------------------------------------------- 
 ################################################################################ 
-  ##### TS-CONFIRMED Cases  -- Data dated:  2021-02-08  ::  2021-02-09 13:02:26 
+  ##### TS-CONFIRMED Cases  -- Data dated:  2021-02-09  ::  2021-02-10 01:02:25 
 ################################################################################ 
   Number of Countries/Regions reported:  192 
   Number of Cities/Provinces reported:  85 
   Unique number of distinct geographical locations combined: 273 
 -------------------------------------------------------------------------------- 
-  Worldwide ts-confirmed  Totals: 106478046 
+  Worldwide ts-confirmed  Totals: 106902907 
 -------------------------------------------------------------------------------- 
-   Country.Region Province.State   Totals GlobalPerc LastDayChange   t-2    t-3    t-7   t-14   t-30
-1              US                27097095      25.45         89727 89581 104015 114437 146597 213281
-2           India                10847304      10.19          9110 11831  12059  11039  12689  16311
-3          Brazil                 9524640       8.95             0 77475      0  54096  61963  29792
-4  United Kingdom                 3959784       3.72         14104 15845  18262  16840  20088  54940
-5          Russia                 3939162       3.70         15701 15808  16379  16406  17982  22540
-6          France                 3327305       3.12          4317 19715  19188  22659  21860  15801
-7           Spain                 2989085       2.81         47095     0      0  29064  36435      0
-8           Italy                 2644707       2.48          7969 11640  13439   9651  10584  18625
-9          Turkey                 2539559       2.39          8103  6670   7897   7795   7103   9138
-10        Germany                 2296323       2.16          4650  6670   8632   7690   9387    948
-11       Colombia                 2161462       2.03          4246  6009   8547  10091  13606  15537
-12      Argentina                 1985501       1.86          5154  3658   6680   9695  10409   7808
-13         Mexico                 1936013       1.82          3868  6065  13209   4384  17165  10003
-14         Poland                 1552686       1.46          2431  4725   5966   4326   4603   9133
-15   South Africa                 1477511       1.39          1376  2435   3184   2649   6041  17421
+   Country.Region Province.State   Totals GlobalPerc LastDayChange   t-2   t-3    t-7   t-14   t-30
+1              US                27189761      25.43         92666 89727 89581 121469 152478 213963
+2           India                10858371      10.16         11067  9110 11831  12899  11666  12584
+3          Brazil                 9599565       8.98         74925     0 77475  56002  63520  25822
+4  United Kingdom                 3972148       3.72         12364 14104 15845  19202  25308  46169
+5          Russia                 3953970       3.70         14808 15701 15808  16222  17464  23018
+6          France                 3345558       3.13         18253  4317 19715  25896  26916   3582
+7           Spain                 3005487       2.81         16402 47095     0  31596  40285  61422
+8           Italy                 2655319       2.48         10612  7969 11640  13182  15191  12530
+9          Turkey                 2548195       2.38          8636  8103  6670   8102   7489  10220
+10        Germany                 2302051       2.15          5728  4650  6670  12487  15636  11706
+11       Colombia                 2166904       2.03          5442  4246  6009  11025  13953  15003
+12      Argentina                 1993295       1.86          7794  5154  3658   9196  10843   8704
+13         Mexico                 1946751       1.82         10738  3868  6065  12153  17944   7594
+14         Poland                 1556685       1.46          3999  2431  4725   6801   6790   4863
+15           Iran                 1481396       1.39          7640  7321  7065   6870   6608   6208
[…]
