+++ man/figures/covid19-SummaryReport.txt
@@ -1,16 +1,16 @@
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ 
 -------------------------------------------------------------------------------- 
 ################################################################################ 
-  ##### TS-CONFIRMED Cases  -- Data dated:  2021-02-09  ::  2021-02-10 01:02:25 
+  ##### TS-CONFIRMED Cases  -- Data dated:  2021-02-09  ::  2021-02-10 13:02:27 
 ################################################################################ 
   Number of Countries/Regions reported:  192 
   Number of Cities/Provinces reported:  85 
   Unique number of distinct geographical locations combined: 273 
 -------------------------------------------------------------------------------- 
-  Worldwide ts-confirmed  Totals: 106902907 
+  Worldwide ts-confirmed  Totals: 106905601 
 -------------------------------------------------------------------------------- 
    Country.Region Province.State   Totals GlobalPerc LastDayChange   t-2   t-3    t-7   t-14   t-30
-1              US                27189761      25.43         92666 89727 89581 121469 152478 213963
+1              US                27192455      25.44         95360 89727 89581 121469 152478 213963
 2           India                10858371      10.16         11067  9110 11831  12899  11666  12584
 3          Brazil                 9599565       8.98         74925     0 77475  56002  63520  25822
 4  United Kingdom                 3972148       3.72         12364 14104 15845  19202  25308  46169
@@ -33,16 +33,16 @@
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ 
 -------------------------------------------------------------------------------- 
 ################################################################################ 
-  ##### TS-DEATHS Cases  -- Data dated:  2021-02-09  ::  2021-02-10 01:02:26 
+  ##### TS-DEATHS Cases  -- Data dated:  2021-02-09  ::  2021-02-10 13:02:27 
 ################################################################################ 
   Number of Countries/Regions reported:  192 
   Number of Cities/Provinces reported:  85 
   Unique number of distinct geographical locations combined: 273 
 -------------------------------------------------------------------------------- 
-  Worldwide ts-deaths  Totals: 2341004 
+  Worldwide ts-deaths  Totals: 2341104 
 -------------------------------------------------------------------------------- 
    Country.Region Province.State Totals Perc LastDayChange  t-2  t-3  t-7 t-14 t-30
-1              US                468103 1.72          3031 1596 1307 3912 3955 2045
+1              US                468203 1.72          3131 1596 1307 3912 3955 2045
 2          Brazil                233520 2.43          1986    0 1500 1254 1283  480
 3          Mexico                168432 8.65          1701  531  414 1707 1623  662
 4           India                155252 1.43            94   78   84  107  123  167
@@ -63,7 +63,7 @@
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ 
 -------------------------------------------------------------------------------- 
 ################################################################################ 
-  ##### TS-RECOVERED Cases  -- Data dated:  2021-02-09  ::  2021-02-10 01:02:26 
+  ##### TS-RECOVERED Cases  -- Data dated:  2021-02-09  ::  2021-02-10 13:02:28 
 ################################################################################ 
[…]
