### 2021-02-11 351668b
daily update of covid19 data/plots/figures: covid19-SummaryReport.txt ... Thu Feb 11 01:03:29 EST 2021-gargantua  [skip ci]
