+++ man/figures/covid19-SummaryReport.txt
@@ -1,30 +1,30 @@
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ 
 -------------------------------------------------------------------------------- 
 ################################################################################ 
-  ##### TS-CONFIRMED Cases  -- Data dated:  2021-02-09  ::  2021-02-10 13:02:27 
+  ##### TS-CONFIRMED Cases  -- Data dated:  2021-02-10  ::  2021-02-11 01:02:28 
 ################################################################################ 
   Number of Countries/Regions reported:  192 
   Number of Cities/Provinces reported:  85 
   Unique number of distinct geographical locations combined: 273 
 -------------------------------------------------------------------------------- 
-  Worldwide ts-confirmed  Totals: 106905601 
+  Worldwide ts-confirmed  Totals: 107339144 
 -------------------------------------------------------------------------------- 
    Country.Region Province.State   Totals GlobalPerc LastDayChange   t-2   t-3    t-7   t-14   t-30
-1              US                27192455      25.44         95360 89727 89581 121469 152478 213963
-2           India                10858371      10.16         11067  9110 11831  12899  11666  12584
-3          Brazil                 9599565       8.98         74925     0 77475  56002  63520  25822
-4  United Kingdom                 3972148       3.72         12364 14104 15845  19202  25308  46169
-5          Russia                 3953970       3.70         14808 15701 15808  16222  17464  23018
-6          France                 3345558       3.13         18253  4317 19715  25896  26916   3582
-7           Spain                 3005487       2.81         16402 47095     0  31596  40285  61422
-8           Italy                 2655319       2.48         10612  7969 11640  13182  15191  12530
-9          Turkey                 2548195       2.38          8636  8103  6670   8102   7489  10220
-10        Germany                 2302051       2.15          5728  4650  6670  12487  15636  11706
-11       Colombia                 2166904       2.03          5442  4246  6009  11025  13953  15003
-12      Argentina                 1993295       1.86          7794  5154  3658   9196  10843   8704
-13         Mexico                 1946751       1.82         10738  3868  6065  12153  17944   7594
-14         Poland                 1556685       1.46          3999  2431  4725   6801   6790   4863
-15           Iran                 1481396       1.39          7640  7321  7065   6870   6608   6208
+1              US                27285621      25.42         93166 95360 89727 123188 168620 226238
+2           India                10871294      10.13         12923 11067  9110  12408  18855  15968
+3          Brazil                 9659167       9.00         59602 74925     0  56873  61811  64025
+4  United Kingdom                 3985161       3.71         13013 12364 14104  20634  28680  45533
+5          Russia                 3968228       3.70         14258 14808 15701  16463  18856  22657
+6          France                 3370645       3.14         25087 18253  4317   -458      0  19267
+7           Spain                 3023601       2.82         18114 16402 47095  29960  34899  25438
+8           Italy                 2668266       2.49         12947 10612  7969  13656  14360  14242
+9          Turkey                 2556837       2.38          8642  8636  8103   7909   7279   9809
+10        Germany                 2311297       2.15          9246  5728  4650  13032  14883  27210
+11       Colombia                 2173347       2.02          6443  5442  4246   9790  12270  14179
+12      Argentina                 2001034       1.86          7739  7794  5154   8891   9471  13783
+13         Mexico                 1957889       1.82         11138 10738  3868      0  18670  14395
+14         Poland                 1563645       1.46          6960  3999  2431   6495   7153   5394
+15           Iran                 1488981       1.39          7585  7640  7321   7040   6527   6408
 -------------------------------------------------------------------------------- 
   Global Perc. Average:  0.37 (sd: 1.8) 
[…]
