+++ man/figures/covid19-SummaryReport.txt
@@ -1,16 +1,16 @@
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ 
 -------------------------------------------------------------------------------- 
 ################################################################################ 
-  ##### TS-CONFIRMED Cases  -- Data dated:  2021-02-10  ::  2021-02-11 01:02:28 
+  ##### TS-CONFIRMED Cases  -- Data dated:  2021-02-10  ::  2021-02-11 13:02:29 
 ################################################################################ 
   Number of Countries/Regions reported:  192 
   Number of Cities/Provinces reported:  85 
   Unique number of distinct geographical locations combined: 273 
 -------------------------------------------------------------------------------- 
-  Worldwide ts-confirmed  Totals: 107339144 
+  Worldwide ts-confirmed  Totals: 107340682 
 -------------------------------------------------------------------------------- 
    Country.Region Province.State   Totals GlobalPerc LastDayChange   t-2   t-3    t-7   t-14   t-30
-1              US                27285621      25.42         93166 95360 89727 123188 168620 226238
+1              US                27287159      25.42         94704 95360 89727 123188 168620 226238
 2           India                10871294      10.13         12923 11067  9110  12408  18855  15968
 3          Brazil                 9659167       9.00         59602 74925     0  56873  61811  64025
 4  United Kingdom                 3985161       3.71         13013 12364 14104  20634  28680  45533
@@ -33,16 +33,16 @@
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ 
 -------------------------------------------------------------------------------- 
 ################################################################################ 
-  ##### TS-DEATHS Cases  -- Data dated:  2021-02-10  ::  2021-02-11 01:02:28 
+  ##### TS-DEATHS Cases  -- Data dated:  2021-02-10  ::  2021-02-11 13:02:29 
 ################################################################################ 
   Number of Countries/Regions reported:  192 
   Number of Cities/Provinces reported:  85 
   Unique number of distinct geographical locations combined: 273 
 -------------------------------------------------------------------------------- 
-  Worldwide ts-deaths  Totals: 2354416 
+  Worldwide ts-deaths  Totals: 2354561 
 -------------------------------------------------------------------------------- 
    Country.Region Province.State Totals Perc LastDayChange  t-2  t-3  t-7 t-14 t-30
-1              US                471422 1.73          3219 3131 1596 3767 4010 4432
+1              US                471567 1.73          3364 3131 1596 3767 4010 4432
 2          Brazil                234850 2.43          1330 1986    0 1232 1386 1110
 3          Mexico                169760 8.67          1328 1701  531    0 1506 1314
 4           India                155360 1.43           108   94   78  120  163  202
@@ -63,7 +63,7 @@
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ 
 -------------------------------------------------------------------------------- 
 ################################################################################ 
-  ##### TS-RECOVERED Cases  -- Data dated:  2021-02-10  ::  2021-02-11 01:02:29 
+  ##### TS-RECOVERED Cases  -- Data dated:  2021-02-10  ::  2021-02-11 13:02:30 
 ################################################################################ 
[…]
