### 2021-02-12 963f7a2
daily update of covid19 data/plots/figures: covid19-SummaryReport.txt ... Fri Feb 12 01:03:29 EST 2021-gargantua  [skip ci]
