### 2021-02-12 #2995 @LogIN-
Hi @fboehm thank you for invitation for the review.

Authors did a good job in developing a package, it can be successfully installed and used.

Here is my opinion on what needs to be changed:

- Authors should remove self-citation to the same article available to preprint server (Ponce, 2020)

When this self-citation is removed authors should describe in more detail functionality of this R package, similar to their preprint publication, otherwise from current paper many aspects of the package will stay unclear.

---
Maybe the best solution is to have peer review of their preprint paper and having that published in JOSS.
