### 2021-02-12 #2995 @mponce0
> @LogIN- thanks for your message. Can you clarify - is the preprint article the same (same title, same text) that is being considered here? Or does it merely have the same title? Thanks again!

The title is the same but it is quite lengthy. When considering submitting the manuscript to JOSS we thought about it but when checking the structure and recommended format we felt that it wouldn't fit with this format.
Here is a link to the pre-print  https://arxiv.org/abs/2009.01091
We will happy to submit the arXiv version instead if that's OK?
