### 2021-02-12 9e87b36
daily update of covid19 data/plots/figures: covid19-SummaryReport.txt ... Fri Feb 12 13:03:20 EST 2021-gargantua  [skip ci]
