### 2021-02-12 #2995 @mponce0
> Hi @fboehm thank you for invitation for the review.
>
> Authors did a good job in developing a package, it can be successfully installed and used.
>
> Here is my opinion on what needs to be changed:
>
> * Authors should remove self-citation to the same article available to preprint server (Ponce, 2020)
>
> When this self-citation is removed authors should describe in more detail functionality of this R package, similar to their preprint publication, otherwise from current paper many aspects of the package will stay unclear.
>
> Maybe the best solution is to have peer review of their preprint paper and having that published in JOSS.

Hi @LogIN- , many thanks for your review and encouraging words.
We just wanted to clarify and expand a little bit on our previous comment to @fboehm , we did not submit the original manuscript from arXiv to JOSS as we thought the format and in particular the length of it wouldn't fit the style of JOSS.
In addition to that, we thought of using the JOSS publication as the main point of references to the "code/package" itself while working in additional publications more aimed to applications and introductory tutorials in how to use the package itself. In particular, considering that we will continue working and adding new features to the package we thought of having a very high level description and reference to the code in this JOSS publication.
As mentioned above, we are OK with submitting the longer manuscript as show in the arXiv version, but even considering previous JOSS publications (eg. https://doi.org/10.21105/joss.02376) we decided to submit this current shorter version of the paper.
