+++ man/figures/covid19-SummaryReport.txt
@@ -1,199 +1,199 @@
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ 
 -------------------------------------------------------------------------------- 
 ################################################################################ 
-  ##### TS-CONFIRMED Cases  -- Data dated:  2021-02-11  ::  2021-02-12 13:02:26 
+  ##### TS-CONFIRMED Cases  -- Data dated:  2021-02-12  ::  2021-02-13 01:02:26 
 ################################################################################ 
   Number of Countries/Regions reported:  192 
   Number of Cities/Provinces reported:  85 
   Unique number of distinct geographical locations combined: 273 
 -------------------------------------------------------------------------------- 
-  Worldwide ts-confirmed  Totals: 107780490 
+  Worldwide ts-confirmed  Totals: 108186991 
 -------------------------------------------------------------------------------- 
-   Country.Region Province.State   Totals GlobalPerc LastDayChange   t-2   t-3    t-7   t-14   t-30
-1              US                27392512      25.42        105353 94704 95360 133558 166113 229871
-2           India                10880603      10.10          9309 12923 11067  11713  13082  16946
-3          Brazil                 9713909       9.01         54742 59602 74925  50872  59826  60899
-4  United Kingdom                 3998655       3.71         13494 13013 12364  19114  29079  47525
-5          Russia                 3983031       3.70         14803 14258 14808  16444  18966  22544
-6          France                 3390952       3.15         20307 25087 18253  45287  46043  23703
-7           Spain                 3041454       2.82         17853 18114 16402  28565  38118  38869
-8           Italy                 2683403       2.49         15137 12947 10612  14213  13563  15773
-9          Turkey                 2564427       2.38          7590  8642  8636   7901   6912   9554
-10        Germany                 2321225       2.15          9928  9246  5728  10835  12831  25566
-11       Colombia                 2179641       2.02          6294  6443  5442   7248  10058  15898
-12      Argentina                 2008345       1.86          7311  7739  7794   8374   9838  12725
-13         Mexico                 1968566       1.83         10677 11138 10738  26626  16374  15873
-14         Poland                 1570658       1.46          7013  6960  3999   6053   6145   9126
-15           Iran                 1496455       1.39          7474  7585  7640   7054   6573   6317
+   Country.Region Province.State   Totals GlobalPerc LastDayChange    t-2   t-3    t-7   t-14   t-30
+1              US                27490037      25.41         97525 105353 94704 104015 142091 234975
+2           India                10892746      10.07         12143   9309 12923  12059  13044  15590
+3          Brazil                 9765455       9.03         51546  54742 59602      0  58462  67758
+4  United Kingdom                 4013799       3.71         15144  13494 13013  18262  23275  48682
+5          Russia                 3997898       3.70         14867  14803 14258  16379  18751  24303
+6          France                 3390952       3.13             0  20307 25087  19188  23857  21228
+7           Spain                 3056035       2.82         14581  17853 18114      0      0  35878
+8           Italy                 2697296       2.49         13893  15137 12947  13439  12713  17243
+9          Turkey                 2572190       2.38          7763   7590  8642   7897   6871   8962
+10        Germany                 2330422       2.15          9197   9928  9246   8632  17518  21343
+11       Colombia                 2185169       2.02          5528   6294  6443   8547   9173  17121
+12      Argentina                 2015496       1.86          7151   7311  7739   6680   6902  13286
+13         Mexico                 1978954       1.83         10388  10677 11138  13209  15337  16468
+14         Poland                 1577036       1.46          6378   7013  6960   5966   5864   9457
+15           Iran                 1503753       1.39          7298   7474  7585   6990   6317   6471
[…]
