+++ man/figures/covid19-SummaryReport.txt
@@ -1,16 +1,16 @@
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ 
 -------------------------------------------------------------------------------- 
 ################################################################################ 
-  ##### TS-CONFIRMED Cases  -- Data dated:  2021-02-12  ::  2021-02-13 01:02:26 
+  ##### TS-CONFIRMED Cases  -- Data dated:  2021-02-12  ::  2021-02-13 13:02:28 
 ################################################################################ 
   Number of Countries/Regions reported:  192 
   Number of Cities/Provinces reported:  85 
   Unique number of distinct geographical locations combined: 273 
 -------------------------------------------------------------------------------- 
-  Worldwide ts-confirmed  Totals: 108186991 
+  Worldwide ts-confirmed  Totals: 108188977 
 -------------------------------------------------------------------------------- 
    Country.Region Province.State   Totals GlobalPerc LastDayChange    t-2   t-3    t-7   t-14   t-30
-1              US                27490037      25.41         97525 105353 94704 104015 142091 234975
+1              US                27492023      25.41         99511 105353 94704 104015 142091 234975
 2           India                10892746      10.07         12143   9309 12923  12059  13044  15590
 3          Brazil                 9765455       9.03         51546  54742 59602      0  58462  67758
 4  United Kingdom                 4013799       3.71         15144  13494 13013  18262  23275  48682
@@ -33,16 +33,16 @@
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ 
 -------------------------------------------------------------------------------- 
 ################################################################################ 
-  ##### TS-DEATHS Cases  -- Data dated:  2021-02-12  ::  2021-02-13 01:02:27 
+  ##### TS-DEATHS Cases  -- Data dated:  2021-02-12  ::  2021-02-13 13:02:29 
 ################################################################################ 
   Number of Countries/Regions reported:  192 
   Number of Cities/Provinces reported:  85 
   Unique number of distinct geographical locations combined: 273 
 -------------------------------------------------------------------------------- 
-  Worldwide ts-deaths  Totals: 2382880 
+  Worldwide ts-deaths  Totals: 2383000 
 -------------------------------------------------------------------------------- 
    Country.Region Province.State Totals Perc LastDayChange  t-2  t-3  t-7 t-14 t-30
-1              US                480767 1.75          5323 3877 3340 2607 2730 3967
+1              US                480887 1.75          5443 3877 3340 2607 2730 3967
 2          Brazil                237489 2.43          1288 1351 1330    0 1279 1131
 3          Mexico                172557 8.72          1323 1474 1328 1496 1495  999
 4           India                155550 1.43           103   87  108   78  127  191
@@ -63,7 +63,7 @@
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ 
 -------------------------------------------------------------------------------- 
 ################################################################################ 
-  ##### TS-RECOVERED Cases  -- Data dated:  2021-02-12  ::  2021-02-13 01:02:27 
+  ##### TS-RECOVERED Cases  -- Data dated:  2021-02-12  ::  2021-02-13 13:02:29 
 ################################################################################ 
[…]
