### 2021-02-13 #2995 @fboehm
Yes, that's right, @mponce0 . The purpose of JOSS is to publish the software and, typically, we prefer short articles that discuss the software.

@LogIN- I think it would be ok, and, possibly helpful, to maintain the citation to the preprint, especially considering that it's not the same as the manuscript that accompanies the software submission to JOSS. I agree, @LogIN-, with you encouraging the authors to detail functionality of the R package, while aiming to keep the JOSS manuscript length at ~ 3 pages or less.

Does that seem reasonable?
