### 2021-02-14 #2995 @mponce0
> Yes, that's right, @mponce0 . The purpose of JOSS is to publish the software and, typically, we prefer short articles that discuss the software.
>
> @LogIN- I think it would be ok, and, possibly helpful, to maintain the citation to the preprint, especially considering that it's not the same as the manuscript that accompanies the software submission to JOSS. I agree, @LogIN-, with you encouraging the authors to detail functionality of the R package, while aiming to keep the JOSS manuscript length at ~ 3 pages or less.
>
> Does that seem reasonable?

I'm certainly more inclined to keep the JOSS paper short, we are already at 5 pages (a bit above 3 without including references).
We will change the title of the pre-print to reflect that arXiv article is the "User's Manual and Examples of Applications" and the JOSS paper as the main reference of the package.

@fboehm  There are two main issues with adding the description for the functions:
 1) it will increase the length of the manuscript (we added some parts due to some at the request from other reviewer),
 2) my main concern is that we will keep adding new functions and capabilities in the future as the situation unfolds and as requested by users (we have already done this a couple of times) and we develop further projects related to the package. Then the description of the functions presented in this paper would turn to be incomplete.

Personally, I think the best would be to use the JOSS paper as the main reference to the package without going into the details of the implementation of the API and allowing users to get the explicit details from the package website and documentation.
As mentioned before, I know other paper at JOSS follow this very same approach.
