### 2021-02-14 #2995 @mponce0
WRT this last point about the functionality of the package, we went back to the manuscript and try to add a few more sentences detailing some of the functionalities at a very "high" level (end of the "Functionalities and Main Features" section).

Please let us know what do you think.
