### 2021-02-17 #2995 @mponce0
> Hi @fboehm & @mponce0,
> sounds reasonable. I would keep exactly the same title for both and as suggested we can use preprint version as more detail reference.. I checked the latest manuscript build and it's clear that preprint contains more information about the package. If you follow this route and this JOSS publication is the main point for the package I guess this will also become "published version" of that preprint article. In this case I am completely fine in recommending this manuscript for publication.

Hi @LogIN- ,
Sure we can do that, we can keep the same title and even more from the arXiv page refer to the JOSS version of the article using the DOI and a "comment" line mentioning that it was published in the JOSS journal.
