### 2021-03-24 #2995 @fboehm
@mponce0 - The references need a few small fixes.

1. Ponce & Sandhel reference needs "R" capitalized. One way to do this is to verify that the bib file has capital R and put curly braces around the letter R. Upon compiling, this should yield a capitalized R in the references section.
2. The url for Smith, et al. reference needs to be fixed.
3. Hornik's article, in the references section, needs a capital R.
4. Health Canada reference - I think that it needs a capital C for Covid19. Please correct me if I'm mistaken on this.

Please make these small fixes. Once they're done, regenerate the pdf and double-check that the fixes worked. Thanks again!
