### 2021-03-24 #2995 @fboehm
I manually checked the DOIs on the downloaded pdf. They all worked and took me to the expected articles.
