### 2021-03-25 #2995 @mponce0
> 👉📄 [Download article proof](https://raw.githubusercontent.com/openjournals/joss-papers/joss.02995/joss.02995/10.21105.joss.02995.pdf) 📄 [View article proof on GitHub](https://github.com/openjournals/joss-papers/blob/joss.02995/joss.02995/10.21105.joss.02995.pdf) 📄 👈

@fboehm the PDF compilation is working now, I think I had a bug in the references file and that's why the compilation was stalling.
