### 2021-03-26 #2995 @mponce0
> Thank you, @mponce0 . Now that your submission is ready for publication, would you please create an archive - on Figshare or Zenodo - and a new release on Github? Please report in this thread the version number for the new release and the doi for the archive. When creating the archive, please be certain to use the same title and author list (and ordering) as in the submission paper.
>
> Here is a link to a tutorial for creating an archive: https://guides.github.com/activities/citable-code/

Many thanks @fboehm
Here is the information requested:
* ver: 2.1.0
* Zenodo DOI: https://doi.org/10.5281/zenodo.4640307
