### 2021-03-27 #2995 @fboehm
@mponce0 - we need the title of the archive to be the same, verbatim, as that of the paper. Please change the title of the archive to: covid19.analytics: An R Package to Obtain, Analyze and Visualize Data from the 2019 Corona Virus Disease Pandemic
