### 2021-03-27 #2995 @mponce0
> @mponce0 we also need the authors list of the paper to match exactly that of the archive. Please fix this.

@fboehm Sorry about that, I thought I have modified these items but I could see that wasn't the case... I think this time around I edited and adjusted title and authors affiliation properly. Can you please confirm this is the case?
