### 2021-03-28 #2995 @mponce0
@fboehm I'm not sure where is this, but at the beginning of "paper.md" in the "literature" branch,
https://raw.githubusercontent.com/mponce0/covid19.analytics/literature/JOSS/paper.md, we have:

---
title: 'covid19.analytics: An R Package to Obtain, Analyze and Visualize Data from the 2019 Coronavirus Disease Pandemic'
tags:
  - R
  - CRAN
  - coronavirus
  - data access
authors:
  - name: Marcelo Ponce^[corresponding author]
    orcid: 0000-0001-5850-7240
    affiliation: "1" # (Multiple affiliations must be quoted)
  - name: Amit Sandhel
    affiliation: 2
affiliations:
 - name: SciNet HPC Consortium, University of Toronto, Canada
   index: 1
 - name: Independent Researcher
   index: 2
date: 21 January 2021
bibliography: references.bib
---

Am I missing this somewhere else?

Many thanks.
