### 2021-03-28 #2995 @mponce0
@fboehm One more thing I forgot to mention before, as one of the reviewer noticed there was a typo in the title of our original submission, ie. instead of "Corona Virus" it should have been "Coronavirus", we fixed this in the manuscript and the Zenodo submission. Can this also be fixed in the JOSS entry?
