### 2021-03-28 #2995 @fboehm
I don't think that we need to fix the JOSS issue title, where you have "Corona Virus", provided that the archive and paper pdf have the correct spelling. I'll double-check with other editors to be sure of this before we do the final acceptance. Thanks for pointing this out.
