### 2021-03-31 #2995 @fboehm
line 8: empowered
line 14: researchers
line 40: As of today
line 45: Moreover, it has been reported that different
line 49: This is especially
line 52: As of today
line 57: no comma
line 63: Awkward sentence. Try "which is available on CRAN."
65: add-ons:

@mponce0 - please fix these typos.
