### 2021-04-03 #2995 @mponce0
> 👋 @openjournals/joss-eics, this paper is ready to be accepted and published.
>
> Check final proof 👉 [openjournals/joss-papers#2177](https://github.com/openjournals/joss-papers/pull/2177)
>
> If the paper PDF and Crossref deposit XML look good in [openjournals/joss-papers#2177](https://github.com/openjournals/joss-papers/pull/2177), then you can now move forward with accepting the submission by compiling again with the flag `deposit=true` e.g.
>
> ```
> @whedon accept deposit=true from branch literature
> ```

@fboehm is this for us to take care of or the editors?
In any case, many thanks for all your help and support, as well as, the referees @JDRomano2 and @LogIN-
