### 2021-04-04 #2995 @danielskatz
For the reference ihaka:1996, please change the URL tag to a DOI tag (removing the https://doi.org/ part)
