+++ JOSS/references.bib
@@ -18,7 +18,7 @@
         Title = "{R: A Language for Data Analysis and Graphics}",
         Volume = 5,
         Year = 1996,
-        url = {https://doi.org/10.1080/10618600.1996.10474713}
+        url = {10.1080/10618600.1996.10474713}
 }
 
 %%%%%%%%%%%%%%%%
