### 2021-04-04 #2995 @danielskatz
I've also created https://github.com/mponce0/covid19.analytics/pull/10 for a bunch of grammar/style issue - let me know what you disagree with, or merge this.
