### 2021-04-04 #2995 @mponce0
> I've also created [mponce0/covid19.analytics#10](https://github.com/mponce0/covid19.analytics/pull/10) for a bunch of grammar/style issue - let me know what you disagree with, or merge this.

Merged! Many thanks for the thorough review and suggestions/edits.
