### 2021-04-04 #2995 @danielskatz
Re ihaka:1996, you didn't change the tag to doi - it's still url
