### 2021-04-05 #2995 @mponce0
> Re ihaka:1996, you didn't change the tag to doi - it's still url

Strange, this was done in this commit
https://github.com/mponce0/covid19.analytics/commit/847d626f57afd49933c2b77916d0b79d3a69b8a6

and the actual version in the repo, https://github.com/mponce0/covid19.analytics/blob/literature/JOSS/references.bib shows the change...
Maybe it got lost from the merge?
