### 2021-04-05 #2995 @danielskatz
I see the `https://doi.org` part was removed, but the tag is `url`, not `doi`
